### 2025-12-06 #9565 @editorialbot
[PRE REVIEW]: VST: A Python-based deep learning tool for segmenting electron microscopy samples
**Submitting author:** @fgdfgfthgr-fox (<a href="http://orcid.org/0009-0007-7171-5029">Yuyao Huang</a>)
**Repository:** https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/
**Branch with paper.md** (empty if default branch): action_joss_paper
**Version:** v0.10.2
