### 2025-12-22 e243a47
Change version number to 0.10.3
