### 2026-01-21 abe8e8a
Update paper to fill the new scope of JOSS
