+++ paper/paper.bib
@@ -202,3 +202,21 @@
   year={2025},
   publisher={Elsevier}
 }
+
+@misc{abernathey2026zarr,
+   author = {Abernathey, Ryan},
+   title = {Zarr},
+   volume = {2026},
+   number = {Jan 20th},
+   url = {https://zarr.dev/},
+   type = {Web Page}
+}
+
+@inproceedings{ronneberger2015u,
+  title={U-net: Convolutional networks for biomedical image segmentation},
+  author={Ronneberger, Olaf and Fischer, Philipp and Brox, Thomas},
+  booktitle={International Conference on Medical image computing and computer-assisted intervention},
+  pages={234--241},
+  year={2015},
+  organization={Springer}
+}
