### 2026-01-25 c3c3667
Use Adaptive Muon optimizer instead
