+++ pl_module.py
@@ -1,4 +1,6 @@
 import math
+import os
+import multiprocessing
 
 import lightning.pytorch as pl
 import torch
@@ -8,14 +10,13 @@ import torch.utils.tensorboard
 
 from Components import DataComponents
 from Components import Metrics
-from Components.AdEMAMix import AdEMAMix
-from adam_atan2_pytorch import AdamAtan2
 from lightning.pytorch.loggers.tensorboard import TensorBoardLogger
 from Networks import *
 from lightning.pytorch.utilities import grad_norm
 from lightning.pytorch.callbacks import LearningRateFinder
 from lightning.pytorch.callbacks import LearningRateMonitor
 from lightning.pytorch.callbacks import StochasticWeightAveraging
+from pytorch_optimizer.optimizer import AdaMuon
 
 
 class FineTuneLearningRateFinder(LearningRateFinder):
@@ -53,6 +54,49 @@ def weight_sequence(alpha, n):
     return sequence
 
 
+def get_parameter_groups_withmuon(model, weight_decay=0.0001):
+    no_decay_keywords = ["bias", "bn", "batch_norm", "layer_norm", "norm"]
+
+    # Create the 4 groups
+    decay_muon_params = []  # decay + muon
+    decay_no_muon_params = []  # decay + no muon
+    no_decay_muon_params = []  # no decay + muon
+    no_decay_no_muon_params = []  # no decay + no muon
+
+    for name, param in model.named_parameters():
+        if not param.requires_grad:
+            continue
+
+        # Determine decay status
+        requires_decay = not any(no_decay_keyword in name for no_decay_keyword in no_decay_keywords)
+
+        # Determine muon status (hidden weights with ndim == 2)
+        is_muon = param.ndim == 2
+
+        # Assign to appropriate group
+        if requires_decay and is_muon:
+            decay_muon_params.append(param)
+        elif requires_decay and not is_muon:
+            decay_no_muon_params.append(param)
+        elif not requires_decay and is_muon:
+            no_decay_muon_params.append(param)
+        else:  # not requires_decay and not is_muon
+            no_decay_no_muon_params.append(param)
+
+    return [
+        # Group 1: decay + muon
+        {'params': decay_muon_params, 'weight_decay': weight_decay, 'use_muon': True},
+
+        # Group 2: decay + no muon
+        {'params': decay_no_muon_params, 'betas': (0.9, 0.95), 'weight_decay': weight_decay, 'use_muon': False},
+
+        # Group 3: no decay + muon
+        {'params': no_decay_muon_params, 'weight_decay': 0.0, 'use_muon': True},
+
+        # Group 4: no decay + no muon
+        {'params': no_decay_no_muon_params, 'betas': (0.9, 0.95), 'weight_decay': 0.0, 'use_muon': False}
+    ]
+
 class PLModule(pl.LightningModule):
     def __init__(self, arch_args, enable_val, enable_mid_visual, mid_visual_image, instance_mode,
                  use_sparse_label_train, use_sparse_label_val, use_sparse_label_test, logging):
@@ -69,7 +113,7 @@ class PLModule(pl.LightningModule):
         self.use_sparse_label_test = use_sparse_label_test
         self.logging = logging
         self.train_metrics, self.val_metrics, self.test_metrics = [], [], []
-        self.lr = 1e-4 # Not the actual LR since it's automatically computed
+        self.lr = 3e-2 # Not the actual LR since it's automatically computed
         self.pixel_ramp_steps = 2048  # Ramp-up the weight of entropy minimisation during the initial 2048 steps
         self.unsupervised_weight = 0.1
         self.p_loss_fn = Metrics.BinaryMetrics("focal")
[…]
