### 2026-01-25 f6a8949
Use Adaptive Muon optimizer instead
