+++ requirements.txt
@@ -16,5 +16,4 @@ numba>=0.59.1
 mrcfile>=1.5.4
 zarr>=3.1.1
 tifffile>=2025.6.11
-adam_atan2_pytorch>=0.1.18
-einops>=0.8.1
+pytorch-optimizer>=3.9.0
