### 2026-01-26 bb6777f
Better AMD performance, install script exit when no ROCm, fix the issue with weight only loading
