+++ README.md
@@ -16,7 +16,7 @@ Volume Segmentation Tool is a python based tool that utilizes deep learning to p
 
 Based on [Pytorch](https://pytorch.org/) backend. Includes a [Gradio](https://www.gradio.app/) based graphical user interface. 
 
-The desired images are  8bit or 16bit grey scale 3D tif/hdf as commonly generated by electron microscopes. The desired labels are either binary mask where 0 is background and 1 is foreground (semantic), or each value represent a different object (instance).
+The desired images are 8bit or 16bit grey scale 3D tif/hdf as commonly generated by electron microscopes. The desired labels are either binary mask where 0 is background and 1 is foreground (semantic), or each value represent a different object (instance).
 
 This tool does __not__ support 2D image segmentation, nor colorful image segmentation. In which case you should consider [ilastik](https://www.ilastik.org/) 
 or [Trainable Weka Segmentation](https://imagej.net/plugins/tws/) or [nnUNet](https://github.com/MIC-DKFZ/nnUNet).
@@ -46,7 +46,6 @@ or [Trainable Weka Segmentation](https://imagej.net/plugins/tws/) or [nnUNet](ht
 
 - Does not support 2D images, nor images with colours (grey scale only)
 - Since it's based on Deep Learning, the tool needs to be used with a discrete GPU
-  - It's possible to run this tool without a discrete GPU, however it will run __extremely__ slowly
   - Recommended minimal GPU requirement: 4GB of Video Memory, made by Nvidia or AMD
 - Since it's based on Deep Learning, the user has to create training samples to train the network before it could perform segmentation
   - For more information, please see [Tutorials](#tutorials).
