+++ paper/paper.md
@@ -56,7 +56,7 @@ Many earlier DL tools developed are highly specific to single sample types, like
 In short, there is a lack of tools that can handle a wide range of VEM data well for generating both semantic and instance segmentation, while at the same time been easy to use, scalable and can be run locally. Which is what motivated us to develop VST - an easy-to-use and adaptive DL tools specifically optimised for generalised VEM image segmentation. 
 
 # Software design
-The core principles of VST lie in the user-friendliness and scalability. The software come with an one-click installer and full documentation on all user-accessible features, aim to enable accessibility for domain experts without machine learning expertise. In terms of scalability, VST uses Zarr[@abernathey2026zarr], a framework for distributed storage, which allows just-in-time, chunked access for datasets much larger than the user's system memory. Which is a common situation within VEM, where datasets of hundreds or thousands of gigabytes scales are present.
+The core principles of VST lie in the user-friendliness and scalability. The software comes with a one-click installer and full documentation on all user-accessible features, with the aim to enable accessibility for domain experts without machine learning expertise. In terms of scalability, VST uses Zarr[@abernathey2026zarr], a framework for distributed storage, which allows just-in-time, chunked access for datasets much larger than the user's system memory. Which is a common situation within VEM, where datasets of hundreds or thousands of gigabytes scales are present.
 
 In terms of design, the software provides a graphical interface over a set of scripts handling various aspects of the deep learning and inference workload. The internal training framework utilises PyTorch [@paszke2019pytorch], the interface compiles terminal commands and activates the Python scripts as needed (Figure 1). For the DL model underneath, VST uses a heavily modified U-Net [@ronneberger2015u], a proven DL architecture for segmentation tasks that is known for its fast convergence and adaptability. The size, depth and other details of the model are configured automatically based on the characteristics of the user's dataset.
 
@@ -84,6 +84,6 @@ The extra section contains two functionalities: exporting the TensorBoard log to
 ChatGPT and DeepSeek were used to generate some Python functions. All generated functions were thoroughly analysed, tested with real-world data, modified and verified to satisfy desired input and output conditions. No generative AI tools were used in the writing of this manuscript, or the preparation of supporting materials.
 
 # Acknowledgements
-We want to thank to: Rhodri Harfoot and Isa de Vries for providing the SARS-CoV-2 infected cells samples, Sai Velamoor and Laura Burga for tumoursphere preparations, Niki Hazelton and Richard Easingwood from the Otago Micro and Nano Imaging centre for data collection, Marina Richena for fruitful discussions. We would also like to thank Bioeconomy Science Ltd for their support during the MSc study and University of Otago Postgraduate Publishing Bursary (Master's) for their support with the time of writing.
+We want to thank to the New Zealand Institute for Bioeconomy Science for their support during the MSc study and University of Otago Postgraduate Publishing Bursary (Master's) for their support with the time of writing.
 
 # References
