### 2026-02-17 #9565 @rich2355
Hi @fgdfgfthgr-fox, I am Richard, a Biostatistician with some deep learning experience, and I will serve as an editor of this paper. While I am searching for reviewers, you also have the rights to suggest the reviewers if you would like to. Thanks!
