### 2026-02-19 #9565 @rich2355
Hi @ibro45, @ClementCaporal, @ductho-le, @RMeli, @ntolley, are you willing to review this submission for JOSS? Thanks!
