### 2026-02-19 #9565 @ClementCaporal
Hello @rich2355, thank you for the proposition. I don't know if a third reviewer is useful for this paper but if needed I can help to review this submission. (I never worked with electron microscopy but I'll try my best to review the code!)

I also have a question regarding the assessment of "novelty", which I hope is appropriate to ask at this stage (sorry if it is not, this is my first time reviewing for JOSS).
@fgdfgfthgr-fox, I noticed you recently published (congratulation!) "A generalist deep-learning volume segmentation tool for volume electron microscopy of biological samples" (September 2025). As I understand it, that work centers on the same tool. Could you clarify what specific new features or improvements we should review for JOSS in "VST" are being submitted here that were not covered in the Journal of Structural Biology publication?

In any case, thank you for your work in the open science software ecosystem!

I hope you have a good day,

Clément
