### 2026-02-19 #10084 @rich2355
Hi @ibro45, @ductho-le, @ClementCaporal, it is so nice that we started the review right after I served as an editor of this submission. @fgdfgfthgr-fox you are expected to communicate with these three reviewers actively for their potential comments/suggestions.

```@editorialbot generate my checklist```

**The JOSS review is different from most other journals.** Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#10084` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. I will also remind the reviewers if there is no reviewing activity for a while (say, one week). However, things are always evolving and please let me know if any of you requires some more time.
