### 2026-02-20 #10084 @rich2355
> [@ClementCaporal](https://github.com/ClementCaporal) Hi Clement, to answer your question in the pre-review stage about the specific new features or improvements. One of the most significant improvements is the ability to support datasets larger than the user's system memory through the use of the Zarr intermediate format. The level of automation has also improved, with more training hyperparameters now being automatically computed. Although not mentioned in the papers, I also made continuous improvements to the robustness of the installation script, as well as to faster and more accurate instance segmentation and incorporating a new optimiser to accelerate model training.
>
> Should I be worried about the missing DOIs warning? I'm pretty sure all the papers I cited are real.

Hi @fgdfgfthgr-fox, any citation, as long as its DOI exists, should have its DOI to be referred in the paper. That said, if something does not have a DOI, feel free to skip it.
