### 2026-02-27 #10084 @ClementCaporal
Rather than making one big review at the end, I will ask small questions while validating the submission in this thread @fgdfgfthgr-fox if it is ok for you, so we can iterate faster. If needed I might open issues directly on your repo.

Concerning **Contribution and authorship:** I notice that the order of the authors is not the same than in Struct. Biol, and that the affiliations are also not the same. (could you double check the changes are correct?). Otherwise, it is clear that the first author made major contributions and I will validate the checkbox

Concerning **Community guidelines**: I think I didn't find in the readme or documentation a section that explicit how the community can "1. Contribute to the software 2. Report issues or problems with the software 3. Seek support". Can you add these info ?

Concerning **Good practices** for python projects/packages:
- The structure of the package. I am not sure about the package structure you followed for a python package. For example adding a [__init__](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/blob/master/__init__.py) at the root of your repo makes the python package at the same level than your readme. I have similar interrogation for the name between `Components`  and  folders `Network/Modules` (and the `*_Components.py` files inside the `Network/Modules` folder). Maybe it is a structure that I don't know about, so if you have ressources about it, I'll be happy to learn. What I usually tend to follow as general python project guidelines are those links: https://docs.python-guide.org/writing/structure/, https://goodresearch.dev/setup.html#create-a-project-skeleton
-  Test and Validation of the package. I think there is no coverage test yet (or I didn't found it?). I would suggest to have at least one script that runs [start_work-flow ](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/blob/2c3be97167331fd0ab38a3921fca569ec5935f8a/workflow.py#L39) with your available image without any UI input and save the model in a `default Models/`. So someone who wants to test your code can check quickly one example without worrying that they missed some parameters in the user interface
- Unused code:  to me unused code  makes it harder for maintenance / understanding, I notice several functions such as `FCDenseNet57`, `FCDenseNet67`, `FCDenseNet103`, all class in `Networks/GANS.py`, `Networks/HighRes3DNets.py` ... it seems that most of the network cannot be used appart from `UNetBasic`, `UNetResidual_Recommended`, `LR`
- Unsused import: some import are never used ([example](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/blob/2c3be97167331fd0ab38a3921fca569ec5935f8a/Networks/Modules/GANS_Components.py#L6C45-L6C58) ) Maybe you can use a linter (for example https://docs.astral.sh/ruff/ ) to check and format your code.

If you think my remarks above for the **Good practice** are relevant, you can make the change in your repo. Otherwise, we can wait for the review of @ibro45, @ductho-le to see if they share my view for python packages.

Concerning **State of the field** : I think the overall paragraph is great and cite important references. I have two tools in mind that also address the need of local user-friendly instance segmentation to train and predict biological-dataset https://biapy.readthedocs.io/en/latest/workflows/instance_segmentation.html (*Conflict of interest, I am a contributor/coauthor*), and https://github.com/HenriquesLab/DL4MicEverywhere

Concerning **Software Design**: In your paper I might merge the section `Software Design` and `The graphical user interface`

I only have a IOs right now, so I have to steal another computer to try the package but the UI already looks nice!

Thank you for your time,
