### 2026-02-28 #10084 @fgdfgfthgr-fox
@ClementCaporal

Contribution and authorship:
Yes, the order has changed, and they are aware of this. In fact, they requested the change in order. Their affiliations are still the same; they are just written differently. For example, I didn't know that Mihnea was affiliated with Okinawa when I was writing the Struct. Biol paper. They have all read the paper and didn't disagree with the orders.

Community guidelines:
I don't get it. Isn't a GitHub repository already a place to "Report issues or problems" or "Seek support"? People can post issues and start discussions there.

Good practices:
1\. This is designed as a tool rather than an installable Python package. I don't know much about package structure though. Maybe I have done something wrong. I will take a look later today or tomorrow.
2\. Good idea — I will try to add a coverage test. Thank you!
3 & 4. Yes, there is lots of unused code for debugging and internal testing. I will clean it up.

State of the field:
Thank you for making me aware of the existance of theses tools. I am not planning to expand on the State of the field section since JOSS has a pretty tight word limit.
