### 2026-02-28 c9d3e30
Switch from U-Net to Swin Transformer! Clean up old codes.
