+++ install_dependencies_Linux.sh
@@ -11,7 +11,7 @@ python_cmd="python3"
 # git executable
 export GIT="git"
 
-# python3 venv without trailing slash (defaults to ${install_dir}/${clone_dir}/venv)
+# python3 venv without trailing slash (defaults to ${install_dir}/venv)
 venv_dir="venv"
 
 ENV_SCRIPT="prepare_environment.py"
@@ -38,7 +38,7 @@ fi
 rocm_version=""
 if [ -f /opt/rocm/.info/version ]; then
     rocm_version=$(cat /opt/rocm/.info/version | cut -d. -f1-2)
-elif [ -f /opt/rocm/.info/version-* ]; then
+elif ls /opt/rocm/.info/version-* 1> /dev/null 2>&1; then
     rocm_version=$(cat /opt/rocm/.info/version-* | cut -d. -f1-2)
 fi
 
@@ -95,14 +95,14 @@ elif [[ -n "$gpu_info" && "$gpu_info" == *"AMD"* ]] || [[ -n "$rocm_version" ]];
         fi
     else
         printf "\e[1m\e[33mWARNING: AMD GPU detected but ROCm installation not found\e[0m\n"
-        exit 1
     fi
 fi
 
 # Set CPU install if no compatible GPU setup found
 if [ $torch_command_set -eq 0 ]; then
-    printf "\e[1m\e[33mInstalling CPU-only version of PyTorch\e[0m\n"
-    export TORCH_COMMAND="pip install torch torchvision"
+    printf "\e[1m\e[31mERROR: No compatible GPU and drivers found.\n"
+    printf "This tool requires a GPU with CUDA or ROCm support. Aborting.\e[0m\n"
+    exit 1
 fi
 
 # Check for venv module
