### 2026-02-28 0f185b7
Add the missing Swin.py and pause mechanism for installation scripts.
