+++ Networks/Modules/Swin.py
@@ -1,0 +1,288 @@
+import torch
+import torch.nn as nn
+from timm.layers import trunc_normal_
+
+def to_3tuple(x):
+    if isinstance(x, (tuple, list)):
+        assert len(x) == 3
+        return tuple(x)
+    return (x, x, x)
+
+
+def window_partition(x, window_size):
+    """
+    Args:
+        x: (B, D, H, W, C)
+        window_size (tuple[int]): window depth, height, width
+
+    Returns:
+        windows: (num_windows*B, wd, wh, ww, C)
+    """
+    B, D, H, W, C = x.shape
+    wd, wh, ww = window_size
+    # Reshape into grid of windows
+    x = x.view(
+        B,  # Batch
+        D // wd, wd,  # Depth windows, window depth
+        H // wh, wh,  # Height windows, window height
+        W // ww, ww,  # Width windows, window width
+        C  # Channels
+    )
+
+    # Rearrange dimensions to bring windows together
+    windows = x.permute(
+        0,  # B
+        1, 3, 5,  # D_windows, H_windows, W_windows
+        2, 4, 6,  # wd, wh, ww (window interior)
+        7  # C
+    ).reshape(-1, wd, wh, ww, C)
+    return windows
+
+
+def window_reverse(windows, window_size, D, H, W):
+    """
+    Args:
+        windows: (num_windows*B, wd, wh, ww, C)
+        window_size (tuple[int]): window depth, height, width
+        D, H, W: original depth, height, width
+
+    Returns:
+        x: (B, D, H, W, C)
+    """
+    wd, wh, ww = window_size
+    # Calculate batch size
+    windows_per_dim = (D // wd, H // wh, W // ww)
+    num_windows_per_image = windows_per_dim[0] * windows_per_dim[1] * windows_per_dim[2]
+    B = windows.shape[0] // num_windows_per_image
+
+    # Reshape windows back to grid
+    x = windows.view(
+        B,  # Batch
+        windows_per_dim[0],  # D_windows
+        windows_per_dim[1],  # H_windows
+        windows_per_dim[2],  # W_windows
+        wd, wh, ww,  # Window interior
+        -1  # Channels
+    )
+
+    # Rearrange and reshape to original dimensions
+    x = x.permute(
+        0,  # B
+        1, 4,  # D_windows, wd
+        2, 5,  # H_windows, wh
+        3, 6,  # W_windows, ww
+        7  # C
+    ).reshape(B, D, H, W, -1)
+    return x
+
+
+class WindowAttention(nn.Module):
+    """
+    3D Window based multi-head self attention with relative position bias.
+    Supports both shifted and non-shifted windows.
+
+    Args:
+        dim (int): Number of input channels.
+        window_size (tuple[int]): Depth, height, width of the window.
+        num_heads (int): Number of attention heads.
+        qkv_bias (bool, optional): If True, add a learnable bias to query, key, value. Default: False
+        qk_scale (float | None, optional): Override default qk scale of head_dim ** -0.5 if set.
+    """
+
+    def __init__(self, dim, window_size, num_heads, qkv_bias=False, qk_scale=None):
+        super().__init__()
+        self.dim = dim
+        self.window_size = window_size  # (wd, wh, ww)
+        self.num_heads = num_heads
+        head_dim = dim // num_heads
+        self.scale = qk_scale or head_dim ** -0.5
+
+        # Relative position bias table: shape (2*wd-1)*(2*wh-1)*(2*ww-1), nH
+        self.relative_position_bias_table = nn.Parameter(
[…]
