+++ workflow.py
@@ -94,6 +94,7 @@ def start_work_flow(args):
         train_loader = DataLoader(dataset=train_dataset, batch_size=args.batch_size,
                                   collate_fn=collate_fn, sampler=sampler,
                                   num_workers=desired_num_workers, persistent_workers=persistent_workers)
+                                  #num_workers=0, persistent_workers=False)
         del train_dataset
         if 'Validation' in args.workflow_box:
             val_dataset = VD(args.val_dataset_path, args.hw_size, args.d_size, instance_mode,
@@ -219,22 +220,23 @@ if __name__ == "__main__":
     multiprocessing.set_start_method('spawn', force=True)
 
     parser = argparse.ArgumentParser(description="Deep Learning Workflow")
-    parser.add_argument("--workflow_box", nargs='+', choices=["Training", "Validation", "Test", "Predict"], default=[],
+    # The default values are for the coverage test using the default data
+    parser.add_argument("--workflow_box", nargs='+', choices=["Training", "Validation", "Test", "Predict"], default=["Training", "Validation", "Predict"],
                         help="Workflows to enable")
     parser.add_argument("--segmentation_mode", choices=["Semantic", "Instance"], default="Semantic",
                         help="Segmentation Mode")
     parser.add_argument("--train_dataset_path", type=str, default="Datasets/train", help="Train Dataset Path")
     parser.add_argument("--augmentation_csv_path", type=str, default="Augmentation Parameters Anisotropic.csv",
                         help="Csv File for Data Augmentation Settings")
-    parser.add_argument("--train_multiplier", type=int, default=8, help="Train Multiplier (Repeats)")
-    parser.add_argument("--batch_size", type=int, default=1, help="Batch Size")
-    parser.add_argument("--pairing_samples", action="store_true", help="Pairing positive and negative samples in a batch")
-    parser.add_argument("--num_epochs", type=int, default=40, help="Number of Epochs")
-    parser.add_argument("--enable_tensorboard", action="store_true", help="Enable TensorBoard Logging")
+    parser.add_argument("--train_multiplier", type=int, default=128, help="Train Multiplier (Repeats)")
+    parser.add_argument("--batch_size", type=int, default=2, help="Batch Size")
+    parser.add_argument("--pairing_samples", default=True, action="store_true", help="Pairing positive and negative samples in a batch")
+    parser.add_argument("--num_epochs", type=int, default=5, help="Number of Epochs") # Five epoch is enough
+    parser.add_argument("--enable_tensorboard", default=True, action="store_true", help="Enable TensorBoard Logging")
     parser.add_argument("--enable_unsupervised", action="store_true", help="Enable Unsupervised Pretraining")
     parser.add_argument("--memory_saving_mode", action="store_true", help="Try save some system memory by dataloading on just single core")
     parser.add_argument("--unsupervised_train_dataset_path", type=str, default="Datasets/unsupervised_train", help="Unsupervised Dataset Path")
-    parser.add_argument("--unsupervised_train_multiplier", type=int, default=64, help="Unsupervised Train Multiplier (Repeats)")
+    parser.add_argument("--unsupervised_train_multiplier", type=int, default=128, help="Unsupervised Train Multiplier (Repeats)")
     parser.add_argument("--tensorboard_path", type=str, default="lightning_logs",
                         help="Path to the folder which the log will be saved to")
     parser.add_argument("--val_dataset_path", type=str, default="Datasets/val", help="Validation Dataset Path")
@@ -243,38 +245,38 @@ if __name__ == "__main__":
     parser.add_argument("--read_existing_model", action="store_true", help="Read Existing Model Weight File")
     parser.add_argument("--existing_model_path", type=str, default="", help="Path to Existing Model Weight File")
[…]
