### 2026-03-01 8e85c31
Fix the Windows coverage script.
