### 2026-03-02 3d8707b
Update readme regarding the new coverage test scripts.
