### 2026-03-02 ac00664
Fix the gradio version and the included image's z-to-xy ratio.
