### 2026-03-02 #10084 @fgdfgfthgr-fox
I have made the necessary changes to VST, including cleaning up the code and adding a coverage test, and updated the relevant documentation.

It is worth noting that during the latest release, VST changed its network backbone from a U-Net-like convolutional network to a Swin Transformer. Should I update my paper to reflect this change?
