### 2026-03-03 #10084 @rich2355
> Thank you [@fgdfgfthgr-fox](https://github.com/fgdfgfthgr-fox)! Yes sorry I wrote a comment and then removed it because I missclicked. Here is it again.
>
> 1. Concerning the U-Net to Swin Transformer I think it is indeed good to relfect the change in your paper, references, and also in your wiki (I think you removed the "Model Architecture" change in the quick walktrough, but you might want to explain in your readme or home wiki which model (and spec) user can use).
> 2. Concerning State of the field: Mentionning other efforts that address a similar research gap you indentifed would strengthen the manuscript because it will help our community to understand how different approaches complement and better contribute. It will also help user with specific needs to navigate the different solutions (In the case of VST, for example a local webui with a Swin-Tensor architecture for VEM is a relevant contribution).  Regarding the word limit, I see from the editorialbot that your paper has "Wordcount for paper.md is 1250", and the limit for JOSS guideline are between 750-1750 (https://joss.readthedocs.io/en/latest/paper.html). or maybe there is another limit I may be overlooking [@rich2355](https://github.com/rich2355)?
> 3. Thank you for the cleaning up of the code. If I am not mistaken, I notice that there are still unused code in `Components/Metrics` (such as DiceLoss, compute_sdf, getvrand...) and in `Network/GANS.py`, `Network/Modules/GANS.py`, `Network/Testing.py`, from your switch to Swin, it seems that`Network/UNet.py` is not used anymore either? And if you remove the code I just mentionned, it seems that the `Modules/General.py` is not used either?
> 4. in case it went under the radar, you can use a linter to remove unused import from your code.
> 5. I notice also that sometimes the code refer to Swin model Transformer and also SwishTransformer. Is it on purpose?
> 6. Concerning testing:
>
> * Sorry I think my first remark was ambiguous. Coverage test is based on unit test: https://bettercodebetterscience.github.io/book/testing/#test-coverage / https://goodresearch.dev/testing#testing-to-maintain-your-sanity. You might want to add unit test to check you are able to I/O tiff files, zarr files.
> * I think that the test you wrote in `coverage_test_*` is useful to check if the installation is working without relying on the graphical interface, but should be renamed to something like `run_example_workflow` or `sanity_check` but not coverage test
>
> 7. Concerning Community Guidelines: Yes Github can be used by the community to report issues or problem. But a code being hosted on github doesn't mean that the community should use github to report issues. See for example https://github.com/torvalds/linux (and their [documentation to report issues](https://github.com/torvalds/linux/blob/master/Documentation/admin-guide/reporting-issues.rst). In your case if you want the community to interact through github, you can state it explicitely either in [CONTRIBUTING.md](https://github.com/tlambert03/microsim/blob/main/CONTRIBUTING.md) or in your [readme](https://github.com/tlambert03/FPbase) for example

1250 words should be fine.
