### 2026-03-04 e5d493f
Fixing the issue with AMP and mid-visualisation being unable to disable
