+++ pl_module_dit.py
@@ -114,7 +114,7 @@ class PLModule(pl.LightningModule):
         param_groups = get_parameter_groups_with_muon(self, weight_decay=0.001)
         optimizer = AdaMuon(param_groups, lr=self.lr, weight_decay=0.001, adamw_lr=3e-4, adamw_wd=0.001)
         scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, mode='min',
-                                                               factor=0.5, patience=50,
+                                                               factor=0.5, patience=30,
                                                                threshold_mode='rel',
                                                                cooldown=0, min_lr=[1e-3, 1e-5, 1e-3, 1e-5])
         metrics = "val_loss" if self.enable_val else "train_loss"
@@ -347,72 +347,70 @@ if __name__ == "__main__":
     #tracemalloc.start()
     #snap1 = tracemalloc.take_snapshot()
     torch.backends.cudnn.enabled = False
-    sizes = [(192, 64, 4)]
-    channel_counts = [4]
-    precisions = ['32']
+    sizes = [(144, 144)]
+    precisions = ['16-mixed']
     batch_sizes = [2]
     for size in sizes:
-        for channel_count in channel_counts:
-            for precision in precisions:
-                for batch_size in batch_sizes:
-                    #predict_dataset = DataComponents.Predict_Dataset("Datasets/predict", 112, 24, 8, 1)
-                    train_dataset = DataComponents.TrainDataset("Datasets/train",
-                                                                "Augmentation Parameters Anisotropic.csv",
-                                                                32,
-                                                                size[0], size[1], True, False, 'default')
-                    '''unsupervised_train_dataset = DataComponents.UnsupervisedDataset("Datasets/unsupervised_train",
-                                                                                    "Augmentation Parameters Anisotropic.csv",
-                                                                                    64,
-                                                                                    size[0], size[1])'''
-                    train_dataset = DataComponents.CollectedDataset(train_dataset, None)
-                    sampler = DataComponents.CollectedSampler(train_dataset, batch_size, None)
-                    collate_fn = DataComponents.custom_collate
-                    train_loader = torch.utils.data.DataLoader(dataset=train_dataset, batch_size=batch_size,
-                                                               collate_fn=collate_fn, sampler=sampler,
-                                                               num_workers=0, pin_memory=True, persistent_workers=False)
-                    # meta_info = predict_dataset.__getmetainfo__()
-                    # predict_loader = torch.utils.data.DataLoader(dataset=predict_dataset, batch_size=1, num_workers=0)
-                    val_dataset = DataComponents.ValDataset("Datasets/val", size[0], size[1], True,
-                                                            "Augmentation Parameters Anisotropic.csv")
-                    val_loader = torch.utils.data.DataLoader(dataset=val_dataset, batch_size=1)
+        for precision in precisions:
+            for batch_size in batch_sizes:
+                #predict_dataset = DataComponents.Predict_Dataset("Datasets/predict", 112, 24, 8, 1)
+                train_dataset = DataComponents.TrainDataset("Datasets/train",
+                                                            "Augmentation Parameters Anisotropic.csv",
+                                                            32,
+                                                            size[0], size[1], False, False, 'default')
+                '''unsupervised_train_dataset = DataComponents.UnsupervisedDataset("Datasets/unsupervised_train",
+                                                                                "Augmentation Parameters Anisotropic.csv",
+                                                                                64,
+                                                                                size[0], size[1])'''
+                train_dataset = DataComponents.CollectedDataset(train_dataset, None)
[…]
