### 2026-03-04 b89f8f5
Update paper about other options, swin transformer and adamuon.
