+++ paper/paper.bib
@@ -228,3 +228,40 @@
   volume={32},
   year={2019}
 }
+
+@inproceedings{liu2021swin,
+  title={Swin transformer: Hierarchical vision transformer using shifted windows},
+  author={Liu, Ze and Lin, Yutong and Cao, Yue and Hu, Han and Wei, Yixuan and Zhang, Zheng and Lin, Stephen and Guo, Baining},
+  booktitle={Proceedings of the IEEE/CVF international conference on computer vision},
+  pages={10012--10022},
+  year={2021}
+}
+
+@article{franco2025biapy,
+  title={BiaPy: accessible deep learning on bioimages},
+  author={Franco-Barranco, Daniel and Andr{\'e}s-San Rom{\'a}n, Jes{\'u}s A and Hidalgo-Cenalmor, Ivan and Backov{\'a}, Lenka and Gonz{\'a}lez-Marfil, Aitor and Caporal, Cl{\'e}ment and Chessel, Anatole and G{\'o}mez-G{\'a}lvez, Pedro and Escudero, Luis M and Wei, Donglai and others},
+  journal={Nature Methods},
+  volume={22},
+  number={6},
+  pages={1124--1126},
+  year={2025},
+  publisher={Nature Publishing Group US New York}
+}
+
+@article{hidalgo2024dl4miceverywhere,
+  title={DL4MicEverywhere: deep learning for microscopy made flexible, shareable and reproducible},
+  author={Hidalgo-Cenalmor, Iv{\'a}n and Pylv{\"a}n{\"a}inen, Joanna W and G. Ferreira, Mariana and Russell, Craig T and Saguy, Alon and Arganda-Carreras, Ignacio and Shechtman, Yoav and Jacquemet, Guillaume and Henriques, Ricardo and others},
+  journal={Nature Methods},
+  volume={21},
+  number={6},
+  pages={925--927},
+  year={2024},
+  publisher={Nature Publishing Group US New York}
+}
+
+@article{si2025adamuon,
+  title={Adamuon: Adaptive muon optimizer},
+  author={Si, Chongjie and Zhang, Debing and Shen, Wei},
+  journal={arXiv preprint arXiv:2507.11005},
+  year={2025}
+}
