### 2026-03-04 e53fd1e
Clean-up code again and update readme to mention how to contribute. Rename converage test to sanity check.
