+++ Components/AdEMAMix.py
@@ -1,123 +1,0 @@
-import math
-import torch
-from torch.optim import Optimizer
-
-
-# https://github.com/nanowell/AdEMAMix-Optimizer-Pytorch
-
-class AdEMAMix(Optimizer):
-    def __init__(self, params, lr=1e-3, betas=(0.9, 0.999, 0.9999), eps=1e-8,
-                 weight_decay=0, alpha=5.0, T_alpha_beta3=None):
-        if not 0.0 <= lr:
-            raise ValueError(f"Invalid learning rate: {lr}")
-        if not 0.0 <= eps:
-            raise ValueError(f"Invalid epsilon value: {eps}")
-        assert len(betas) == 3, f"Invalid beta parameters: {betas}, expected 3"
-        assert all(0.0 <= beta < 1.0 for beta in betas), f"Invalid beta parameters: {betas}"
-        if not 0.0 <= weight_decay:
-            raise ValueError(f"Invalid weight_decay value: {weight_decay}")
-
-        defaults = dict(lr=lr, betas=betas, eps=eps, weight_decay=weight_decay,
-                        alpha=alpha, T_alpha_beta3=T_alpha_beta3)
-        super(AdEMAMix, self).__init__(params, defaults)
-
-    def __setstate__(self, state):
-        super(AdEMAMix, self).__setstate__(state)
-
-    @torch.no_grad()
-    def step(self, closure=None):
-        loss = None
-        if closure is not None:
-            with torch.enable_grad():
-                loss = closure()
-
-        for group in self.param_groups:
-            params_with_grad = []
-            grads = []
-            exp_avgs = []
-            exp_avg_sqs = []
-            exp_avg_slow = []
-            state_steps = []
-
-            for p in group['params']:
-                if p.grad is not None:
-                    params_with_grad.append(p)
-                    if p.grad.is_sparse:
-                        raise RuntimeError('AdEMAMix does not support sparse gradients')
-                    grads.append(p.grad)
-
-                    state = self.state[p]
-                    # Lazy state initialization
-                    if len(state) == 0:
-                        state['step'] = 0
-                        # Exponential moving average of gradient values
-                        state['exp_avg'] = torch.zeros_like(p, memory_format=torch.preserve_format)
-                        # Exponential moving average of squared gradient values
-                        state['exp_avg_sq'] = torch.zeros_like(p, memory_format=torch.preserve_format)
-                        # Slow exponential moving average
-                        state['exp_avg_slow'] = torch.zeros_like(p, memory_format=torch.preserve_format)
-
-                    exp_avgs.append(state['exp_avg'])
-                    exp_avg_sqs.append(state['exp_avg_sq'])
-                    exp_avg_slow.append(state['exp_avg_slow'])
-                    state['step'] += 1
-                    state_steps.append(state['step'])
-
-            beta1, beta2, beta3 = group['betas']
-            alpha = group['alpha']
-            T_alpha_beta3 = group['T_alpha_beta3']
-
-            self._update_adamemix(
-                params_with_grad,
-                grads,
-                exp_avgs,
-                exp_avg_sqs,
-                exp_avg_slow,
-                state_steps,
-                beta1=beta1,
-                beta2=beta2,
-                beta3=beta3,
-                alpha=alpha,
-                T_alpha_beta3=T_alpha_beta3,
-                lr=group['lr'],
-                weight_decay=group['weight_decay'],
-                eps=group['eps'],
-            )
-
-        return loss
-
[…]
