+++ Networks/Modules/Transformer.py
@@ -1,163 +1,0 @@
-import torch
-import torch.nn as nn
-
-class RotaryEmbedding3D(nn.Module):
-    def __init__(self, dim, freq=50):
-        super().__init__()
-        assert dim % 3 == 0, "head dim must be divisible by 3 for 3D rotary embedding"
-        self.dim = dim
-        self.third_dim = dim // 3
-
-        # Create frequency tensors for depth, height, and width
-        inv_freq = 1.0 / (freq ** (torch.arange(0, self.third_dim, 1).float() / self.third_dim))
-        self.register_buffer("inv_freq", inv_freq)
-
-
-    def reshape(self, x):
-        return x.reshape(-1, x.shape[-1])
-
-    def forward(self, grid_size):
-        D, H, W = grid_size
-        device = self.inv_freq.device
-
-        # Create grid coordinates
-        d_pos = torch.arange(D, device=device, dtype=self.inv_freq.dtype)
-        h_pos = torch.arange(H, device=device, dtype=self.inv_freq.dtype)
-        w_pos = torch.arange(W, device=device, dtype=self.inv_freq.dtype)
-
-        # Compute sinusoidal embeddings for depth, height, and width separately
-        # Each will have shape: (position, third_dim)
-        sin_d = torch.sin(d_pos[:, None] * self.inv_freq[None, :])
-        cos_d = torch.cos(d_pos[:, None] * self.inv_freq[None, :])
-        sin_h = torch.sin(h_pos[:, None] * self.inv_freq[None, :])
-        cos_h = torch.cos(h_pos[:, None] * self.inv_freq[None, :])
-        sin_w = torch.sin(w_pos[:, None] * self.inv_freq[None, :])
-        cos_w = torch.cos(w_pos[:, None] * self.inv_freq[None, :])
-
-        # Expand to full 3D grid
-        # For depth: (D, H, W, third_dim)
-        sin_d = sin_d.unsqueeze(1).unsqueeze(1).repeat(1, H, W, 1)
-        cos_d = cos_d.unsqueeze(1).unsqueeze(1).repeat(1, H, W, 1)
-        # For height: (D, H, W, third_dim)
-        sin_h = sin_h.unsqueeze(0).unsqueeze(2).repeat(D, 1, W, 1)
-        cos_h = cos_h.unsqueeze(0).unsqueeze(2).repeat(D, 1, W, 1)
-        # For width: (D, H, W, third_dim)
-        sin_w = sin_w.unsqueeze(0).unsqueeze(1).repeat(D, H, 1, 1)
-        cos_w = cos_w.unsqueeze(0).unsqueeze(1).repeat(D, H, 1, 1)
-
-        # Flatten spatial dimensions
-        sin_d = self.reshape(sin_d)
-        cos_d = self.reshape(cos_d)
-        sin_h = self.reshape(sin_h)
-        cos_h = self.reshape(cos_h)
-        sin_w = self.reshape(sin_w)
-        cos_w = self.reshape(cos_w)
-
-        return (sin_d, cos_d), (sin_h, cos_h), (sin_w, cos_w)
-
-
-def rotate_half(x):
-    """Rotates half the hidden dims of the input."""
-    x1, x2 = x.chunk(2, dim=-1)
-    return torch.cat((-x2, x1), dim=-1)
-
-
-def apply_rotary_pos_emb_3d(x, sin_d, cos_d, sin_h, cos_h, sin_w, cos_w):
-    """
-    Apply rotary embedding by splitting x into three thirds:
-    - First third rotated based on depth (z position)
-    - Second third rotated based on height (y position)
[…]
