+++ paper/paper.md
@@ -55,12 +55,12 @@ Many earlier DL tools developed are highly specific to single sample types, like
 
 In short, there is a lack of tools that can handle a wide range of VEM data well for generating both semantic and instance segmentation, while at the same time been easy to use, scalable and can be run locally. Which is what motivated us to develop VST - an easy-to-use and adaptive DL tools specifically optimised for generalised VEM image segmentation.
 
-When VST was developed, BiaPy [franco2025biapy] and DL4MicEverywhere [hidalgo2024dl4miceverywhere] emerged as competitive options that satisfied the above criteria. They share features such as a GUI and semantic and instance segmentation on 3D images. Apart from not relying on Docker and specialising solely in 3D semantic and instance segmentation, VST incorporates more recent and sophisticated deep learning techniques to improve speed and accuracy (see the sections below).
+When VST was developed, BiaPy [@franco2025biapy] and DL4MicEverywhere [@hidalgo2024dl4miceverywhere] emerged as competitive options that satisfied the above criteria. They share features such as a GUI and semantic and instance segmentation on 3D images. Apart from not relying on Docker and specialising solely in 3D semantic and instance segmentation, VST incorporates more recent and sophisticated deep learning techniques to improve speed and accuracy (see the sections below).
 
 # Software design
 The core principles of VST lie in the user-friendliness and scalability. The software comes with a one-click installer and full documentation on all user-accessible features, with the aim to enable accessibility for domain experts without machine learning expertise. In terms of scalability, VST uses Zarr[@abernathey2026zarr], a framework for distributed storage, which allows just-in-time, chunked access for datasets much larger than the user's system memory. Which is a common situation within VEM, where datasets of hundreds or thousands of gigabytes scales are present.
 
-In terms of design, the software provides a graphical interface over a set of scripts handling various aspects of the deep learning and inference workload. The internal training framework utilises PyTorch [@paszke2019pytorch], the interface compiles terminal commands and activates the Python scripts as needed (Figure 1). For the DL model underneath, VST uses a 3D version of Swin transformer [@liu2021swin] for the DL model underneath. This is a proven DL architecture for a wide range of vision tasks and is known for its superior performance to the traditional U-Net [@ronneberger2015u] architecture, as well as its linear computational complexity. The size, depth, and other details of the model are configured automatically based on the characteristics of the user's dataset. Another recent technique that VST has integrated is the use of adaptive muon [si2025adamuon] as the optimiser due to its faster convergence.
+In terms of design, the software provides a graphical interface over a set of scripts handling various aspects of the deep learning and inference workload. The internal training framework utilises PyTorch [@paszke2019pytorch], the interface compiles terminal commands and activates the Python scripts as needed (Figure 1). For the DL model underneath, VST uses a 3D version of Swin transformer [@liu2021swin] for the DL model underneath. This is a proven DL architecture for a wide range of vision tasks and is known for its superior performance to the traditional U-Net [@ronneberger2015u] architecture, as well as its linear computational complexity. The size, depth, and other details of the model are configured automatically based on the characteristics of the user's dataset. Another recent technique that VST has integrated is the use of adaptive muon [@si2025adamuon] as the optimiser due to its faster convergence.
 
 Much of VST's internal logic is optimised for single-class semantic and instance segmentation and cannot easily be transferred to the multi-class case or 2D image segmentation. This trade-off was made to keep workloads manageable, simplify the codebase, and maximise the ease of use for those with minimal machine learning expertise.
 
