### 2026-03-04 #10084 @fgdfgfthgr-fox
@ClementCaporal
Hi there,
1. I have made changes to the paper to reflect the fact that the VST now operates on the Swin transformer and the adaptive muon optimiser.
2. I had a look at them and added another paragraph to the "Statement of need" section regarding the two options you mentioned.
3. I have made further clean-up changes, so VST should now be free of unused files.
4. I use PyCharm which has a similar function. No worries.
5. Sorry! I misspelled! It's all "swin" now.
6. I renamed them to sanity_check instead. Not going to add unit test, though, that seems unnecessary for this software.
7. I added a line to the readme file to tell contributors to use GitHub for feedback. I am still working on a more rigorous evaluation of the new Swin Transformer's performance, so it will take some time before I can add specification recommendations, etc. One thing I am certain of is that it is going to outperform nnU-Net by a large margin...
