### 2026-03-04 #10084 @ClementCaporal
Hello @fgdfgfthgr-fox thank you for your quick answers, and the time you take to answer each of my points.

Some minor remark:
1. It seems that your 3D implementation is an adaptation from https://github.com/huggingface/pytorch-image-models/blob/f4f4cfc73f3e53b1f428f1abd364cfbf41323dd3/timm/models/swin_transformer.py#L255 . If it is the case, you might cite their [repo](https://github.com/huggingface/pytorch-image-models?tab=readme-ov-file#citing) in your **Software design** to respect their Apache License 2.0 ?
2. In the new paper version I noticed this repetition "For the DL model underneath, VST uses a 3D version
of Swin transformer (Liu et al., 2021) for the DL model underneat"
3. Following the JOSS structure, you could split your **Statement of need** as **Statement of need** and  **State of the field**
3. In your readme some remark:
- update the coverage test name to sanity_test in your README
- "Since it's based on Deep Learning, the user has to create training samples to train the network before it could perform segmentation " I am not sure I understand your sentence here. Do you mean the user needs to create annotation/ground truth because the model is supervised during training?
