### 2026-03-04 #10084 @ClementCaporal
Hello @rich2355 , the repo had a quite substantial update from the first `@editorialbot check repository` I would like to get a new report for the new main branch state. Is it something the authors should do by merging themself the branches or is it something to do from the editorialbot ?
Thank you!
