### 2026-03-05 3f5b30d
Add the missing padding to cropping when doing zarr. Update readme.
