+++ paper/paper.bib
@@ -265,3 +265,13 @@
   journal={arXiv preprint arXiv:2507.11005},
   year={2025}
 }
+
+@misc{rw2019timm,
+  author = {Ross Wightman},
+  title = {PyTorch Image Models},
+  year = {2019},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  doi = {10.5281/zenodo.4414861},
+  howpublished = {\url{https://github.com/rwightman/pytorch-image-models}}
+}
