### 2026-03-05 #10084 @fgdfgfthgr-fox
@rich2355
Hi, not sure what I need to do. I see that ClementCaporal has already @ed the editorialbot and I am not sure what else is needed.
@ClementCaporal
OK, I've made the minor changes you mentioned, so hopefully it's ready to go.
