### 2026-03-05 #10084 @ClementCaporal
@fgdfgfthgr-fox, when I use `@editorialbot check repository` it will fetch the repo branch [`action_joss_paper`](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/tree/action_joss_paper)  which is not up to date with your repo main branch. So the stats are not up to date. To solve this you need to merge `main` on `action_joss_paper`

Thank you for the changes!
