### 2026-03-05 #10084 @rich2355
> [@rich2355](https://github.com/rich2355) Hi, not sure what I need to do. I see that ClementCaporal has already [@ed](https://github.com/ed) the editorialbot and I am not sure what else is needed. [@ClementCaporal](https://github.com/ClementCaporal) OK, I've made the minor changes you mentioned, so hopefully it's ready to go.

Hi @fgdfgfthgr-fox, just a kind reminder to not to at someone who is not supposed to belong to this thread. I am sure this is probably just a mistype but please double-check next time. Thanks!
