### 2026-03-07 #10084 @ductho-le
Hi @fgdfgfthgr-fox, thank you for your work on VST. I've completed my review and opened three issues on the repository for the major items:

[#4 — Add automated test suite and CI pipeline](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/issues/4)
[#5 — Add CONTRIBUTING.md and CODE_OF_CONDUCT.md](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/issues/5)
[#6 — Fix BibTeX bibliography formatting and missing citations](https://github.com/fgdfgfthgr-fox/Volume_Seg_Tool/issues/6)

A few additional minor items in paper.md (checked against the action_joss_paper branch):

**Grammar fixes:**

"...while at the same time been easy to use" → should be "being easy"
"Which is what motivated us to develop VST" → sentence fragment, suggest "This is what motivated us..." or "This gap motivated us..."
"...uses Zarr[@abernathey2026zarr]" → missing space before citation: "Zarr [@abernathey2026zarr]"
"Which is a common situation within VEM" → another sentence fragment, suggest "This is a common situation..."

**Benchmarks**: The README notes evaluation is underway after the Swin Transformer switch — once complete, consider adding reproducibility scripts or linking to the specific figures/tables in the companion paper.
