### 2026-03-09 5319447
Create python-app-linux.yml
CI for Linux
