+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+# Contributing to VST
+Well, thank you for wanting to contribute to VST. Whether it's by reporting a bug, suggesting new features or even open a pull request. 
+This documentation provides a high-level overview of how you can (and should) do each of these things.
+
+## Submitting an Issue
+Are you getting very bad results for your dataset? 
+Did the installation scripts fail?
+Did VST throw an error message even though you think you have done everything correctly? 
+Has your computer frozen (or even crashed) when running VST?
+
+Well, unfortunate to hear that. Take a look at the existing issues.
+If your issue already exists, you can give it a thumbs-up.
+
+If not, you should open a new issue. When describing your issue, please provide as much detail as possible about what happened. 
+You should also include your Python version, which operating system you are using, and your GPU model.
+
+## Suggesting New Features
+The Limitations section of the readme file states what VST will not be.
+It's unlikely that I will try to implement anything that is listed there.
+
+Other than that, please submit an issue, state that it is a new feature request and explain why you need that feature.
+
+## Pull Request
+So far, VST has been almost entirely written by me alone. 
+My programming skills are not just self-taught; I was still constantly learning new things until recently. 
+Therefore, VST's code is likely to be very difficult for others to understand. 
+I would like to thank everyone who has contributed by writing pull requests for VST.
+
+I don't have much to say, other than please be detailed on your description on what this pull request is for and what does it adds.
+
+## Email Me
+Though unlikely, if you find any security flaw in VST that would be dangerous to put on as a GitHub issue, 
+you can email me at yuyaoh6@gmail.com
