+++ paper/paper.bib
@@ -1,9 +1,14 @@
-@article{abid2019gradio,
-  title={Gradio: Hassle-free sharing and testing of ml models in the wild},
-  author={Abid, Abubakar and Abdalla, Ali and Abid, Ali and Khan, Dawood and Alfozan, Abdulrahman and Zou, James},
-  journal={arXiv preprint arXiv:1906.02569},
-  year={2019}
-}
+@article{AbubakarAbid2019GHSa,
+author={Abid,Abubakar and Abdalla,Ali and Abid,Ali and Khan,Dawood and Alfozan,Abdulrahman and Zou,James},
+year={2019},
+title={Gradio: Hassle-Free Sharing and Testing of ML Models in the Wild},
+journal={arXiv.org},
+note={Copyright - © 2019. This work is published under http://arxiv.org/licenses/nonexclusive-distrib/1.0/ (the “License”). Notwithstanding the ProQuest Terms and Conditions, you may use this content in accordance with the terms of the License; Last updated - 2023-02-02},
+abstract={Accessibility is a major challenge of machine learning (ML). Typical ML models are built by specialists and require specialized hardware/software as well as ML experience to validate. This makes it challenging for non-technical collaborators and endpoint users (e.g. physicians) to easily provide feedback on model development and to gain trust in ML. The accessibility challenge also makes collaboration more difficult and limits the ML researcher's exposure to realistic data and scenarios that occur in the wild. To improve accessibility and facilitate collaboration, we developed an open-source Python package, Gradio, which allows researchers to rapidly generate a visual interface for their ML models. Gradio makes accessing any ML model as easy as sharing a URL. Our development of Gradio is informed by interviews with a number of machine learning researchers who participate in interdisciplinary collaborations. Their feedback identified that Gradio should support a variety of interfaces and frameworks, allow for easy sharing of the interface, allow for input manipulation and interactive inference by the domain expert, as well as allow embedding the interface in iPython notebooks. We developed these features and carried out a case study to understand Gradio's usefulness and usability in the setting of a machine learning collaboration between a researcher and a cardiologist.},
+keywords={Business And Economics--Banking And Finance; Machine Learning; Human-Computer Interaction; Researchers; Accessibility; Source code; Collaboration; Feedback; Physicians; Freeware; Artificial intelligence},
+language={English},
+url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-ml-models-wild/docview/2236498323/se-2},
+} 
 
 @article{arganda2017trainable,
   title={Trainable Weka Segmentation: a machine learning tool for microscopy pixel classification},
@@ -203,13 +208,50 @@
   publisher={Elsevier}
 }
 
-@misc{abernathey2026zarr,
-   author = {Abernathey, Ryan},
-   title = {Zarr},
-   volume = {2026},
-   number = {Jan 20th},
-   url = {https://zarr.dev/},
-   type = {Web Page}
+@software{falcon_wa_2020_3773450,
+  author={Falcon, WA},
+  title={PyTorch Lightning},
+  month=march,
+  year=2019,
+  version={1.4},
+  doi={10.5281/zenodo.3828935},
+  url={https://www.pytorchlightning.ai},
+}
+
+@software{alistair_miles_2020_3828935,
+  author={Alistair Miles and
+                  John Kirkham and
+                  Martin Durant and
+                  James Bourbeau and
+                  Tarik Onalan and
+                  Joe Hamman and
+                  Zain Patel and
+                  shikharsg and
+                  Matthew Rocklin and
+                  raphael dussin and
+                  Vincent Schut and
+                  Elliott Sales de Andrade and
[…]
