### 2026-04-01 #10084 @fgdfgfthgr-fox
Could I ask how things have been going?
