### 2026-04-07 #10084 @rich2355
> Could I ask how things have been going?

Thanks for the reminder! I am close to recover from a long travel.

Hi all, I am simply checking the review status (again). It seems that @ClementCaporal and @ductho-le have provided suggestions and the author's are actively addressing them. I would like to confirm whether @ClementCaporal has finished the review, and whether @ductho-le is ok with the @fgdfgfthgr-fox's most recent response?

In addition, @ibro45 I find you have not yet started your review. If you are not expected to start reviewing in the next few weeks, please feel free to let me know so that I can start finding other reviewers to move forward.

Thanks all!
