### 2026-05-04 c48885c
Small optimisations on fold and unfold functions.
