+++ Networks/DiT.py
@@ -1,8 +1,63 @@
 import torch
-import unfoldNd
 import torch.nn as nn
 from torch.utils.checkpoint import checkpoint
 from .Modules.Swin import SwinTransformerBlock, compute_attn_mask
+
+def unfold(x, kernel_size):
+    """
+    Unfold a 3D tensor (B, C, D, H, W) using a non‑overlapping 3D sliding window.
+    Returns:
+        patches: Tensor of shape (B, num_patches, C * kd * kh * kw)
+                 where num_patches = (D//kd) * (H//kh) * (W//kw)
+    """
+    B, C, D, H, W = x.shape
+    kd, kh, kw = kernel_size
+
+    x = x.reshape(B, C, D // kd, kd, H // kh, kh, W // kw, kw)
+
+    # Permute to bring block dimensions together: (B, C, Db, Hb, Wb, kd, kh, kw)
+    # where Db = D//kd, etc.
+    x = x.permute(0, 1, 2, 4, 6, 3, 5, 7)
+
+    # Flatten kernel dimensions: (B, C, Db*Hb*Wb, kd*kh*kw)
+    num_patches = (D // kd) * (H // kh) * (W // kw)
+    x = x.reshape(B, C, num_patches, -1)
+
+    # Move the channel dimension to the end: (B, num_patches, C*kd*kh*kw)
+    x = x.permute(0, 2, 1, 3).reshape(B, num_patches, -1)
+
+    return x
+
+def fold(patches, kernel_size, original_shape):
+    """
+    Fold patches back into a 3D tensor (B, C, D, H, W).
+    Assumes non‑overlapping windows where kernel size equals stride.
+
+    Args:
+        patches: Tensor of shape (B, num_patches, C * kd * kh * kw)
+        kernel_size: tuple (kd, kh, kw)
+        original_shape: tuple (B, C, D, H, W) of the original tensor
+
+    Returns:
+        x: Tensor of shape (B, C, D, H, W)
+    """
+    B, C, D, H, W = original_shape
+    kd, kh, kw = kernel_size
+
+    Db, Hb, Wb = D // kd, H // kh, W // kw  # number of blocks per dimension
+    num_patches = Db * Hb * Wb
+
+    # Reshape to separate channels from kernel product
+    patches = patches.reshape(B, num_patches, C, kd * kh * kw)
+    # Bring channels forward
+    patches = patches.permute(0, 2, 1, 3)               # (B, C, num_patches, kd*kh*kw)
+    # Split num_patches into block indices and kernel product into spatial dims
+    patches = patches.reshape(B, C, Db, Hb, Wb, kd, kh, kw)
+    # Interleave block and kernel dimensions (reverse of unfold's permutation)
+    patches = patches.permute(0, 1, 2, 5, 3, 6, 4, 7)   # (B, C, Db, kd, Hb, kh, Wb, kw)
+    # Merge to full spatial dimensions
+    return patches.reshape(B, C, D, H, W)
+
 
 class SwinBlock(nn.Module):
     def __init__(self, d_main, nhead, num_layers, ffn_multiplier, window_size):
@@ -32,7 +87,7 @@ class Network(nn.Module):
         self.window_size = window_size
         self.instance = instance
         patch_dim = (self.patch_size[0]*self.patch_size[1]*self.patch_size[2])
-        self.to_patches = unfoldNd.UnfoldNd(kernel_size=self.patch_size, stride=self.patch_size)
+        #self.to_patches = unfoldNd.UnfoldNd(kernel_size=self.patch_size, stride=self.patch_size)
         self.up = nn.Linear(patch_dim, patch_dim*2)
         self.dit = SwinBlock(patch_dim*2, 2, depth, 2, self.window_size)
[…]
