+++ requirements.txt
@@ -17,6 +17,5 @@ mrcfile>=1.5.4
 zarr>=3.1.1
 tifffile>=2025.6.11
 pytorch-optimizer>=3.9.0
-unfoldNd>=0.2.3
 timm>=1.0.24
 pytest>=9.0.2
