+++ pl_module_simmim.py
@@ -1,0 +1,168 @@
+import math
+
+import lightning.pytorch as pl
+import torch
+import torch.utils.data
+import torch.utils.tensorboard
+
+from Components import DataComponents
+from Components import SimMIM
+from lightning.pytorch.loggers.tensorboard import TensorBoardLogger
+from Networks import *
+from lightning.pytorch.callbacks import LearningRateMonitor
+from pytorch_optimizer.optimizer import AdaMuon
+
+
+device = "cuda" if torch.cuda.is_available() else "cpu"
+
+
+def get_parameter_groups_with_muon(model, weight_decay=0.0001):
+    no_decay_keywords = ["bias", "bn", "batch_norm", "layer_norm", "norm", "RMSNorm"]
+
+    # Create the 4 groups
+    decay_muon_params = []  # decay + muon
+    decay_no_muon_params = []  # decay + no muon
+    no_decay_muon_params = []  # no decay + muon
+    no_decay_no_muon_params = []  # no decay + no muon
+
+    for name, param in model.named_parameters():
+        if not param.requires_grad:
+            continue
+
+        # Determine decay status
+        requires_decay = not any(no_decay_keyword in name for no_decay_keyword in no_decay_keywords)
+
+        # Determine muon status (hidden weights with ndim == 2)
+        is_muon = param.ndim == 2
+
+        # Assign to appropriate group
+        if requires_decay and is_muon:
+            decay_muon_params.append(param)
+        elif requires_decay and not is_muon:
+            decay_no_muon_params.append(param)
+        elif not requires_decay and is_muon:
+            no_decay_muon_params.append(param)
+        else:  # not requires_decay and not is_muon
+            no_decay_no_muon_params.append(param)
+
+    return [
+        # Group 1: decay + muon
+        {'params': decay_muon_params, 'weight_decay': weight_decay, 'use_muon': True},
+
+        # Group 2: decay + no muon
+        {'params': decay_no_muon_params, 'betas': (0.9, 0.95), 'weight_decay': weight_decay, 'use_muon': False},
+
+        # Group 3: no decay + muon
+        {'params': no_decay_muon_params, 'weight_decay': 0.0, 'use_muon': True},
+
+        # Group 4: no decay + no muon
+        {'params': no_decay_no_muon_params, 'betas': (0.9, 0.95), 'weight_decay': 0.0, 'use_muon': False}
+    ]
+
+class PLModule(pl.LightningModule):
+    def __init__(self, arch_args, enable_mid_visual, logging):
+        super().__init__()
+        self.save_hyperparameters()
+        self.patch_dim = arch_args[0]
+        self.network = DiT.SwinTransformerForSimMIM(*arch_args)
+        self.mim_loss_fn = DiT.SimMIM(self.network)
+        self.enable_mid_visual = enable_mid_visual
+        self.logging = logging
+        self.train_metrics, self.val_metrics, self.test_metrics = [], [], []
+        self.lr = 3e-2
+
+    def forward(self, image, mask):
+        return self.mim_loss_fn(image, mask)
+
+    def configure_optimizers(self):
+        #fused = True if device == "cuda" else False
+        param_groups = get_parameter_groups_with_muon(self, weight_decay=0.001)
+        optimizer = AdaMuon(param_groups, lr=self.lr, weight_decay=0.001, adamw_lr=3e-4, adamw_wd=0.001)
+        scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer, mode='min',
+                                                               factor=0.5, patience=30,
+                                                               threshold_mode='rel',
+                                                               cooldown=0, min_lr=[1e-3, 1e-5, 1e-3, 1e-5])
+        metrics = "train_loss"
[…]
