### 2026-05-22 36e19d9
Update required PyTorch Version, improve the Swin architecture for better accuracy, allow prediction file to be larger than 4GB, automatic mid-visualisation
