+++ requirements.txt
@@ -16,6 +16,7 @@ numba>=0.59.1
 mrcfile>=1.5.4
 zarr>=3.1.1
 tifffile>=2025.6.11
-pytorch-optimizer>=3.9.0
+pytorch-optimizer>=3.10.0
 timm>=1.0.24
 pytest>=9.0.2
+aule-attention>=0.5.0

[…]
