### 2026-05-22 eaae09f
Change default xy patch size to 6 so coverage test actually works.
