### 2026-05-22 91f48b1
Remove reference to mid_visualisation_input
