### 2026-05-22 8bc52aa
Enable Mid Visualisation by default
