### 2026-05-22 be8c73d
Change version number for 0.12.0 release
