### 2026-05-23 #10084 @fgdfgfthgr-fox
@ClementCaporal @rich2355
I have resolved the issues with the citations.
Google Colab is impossible. They have a maximum time limit, and nnUNet takes far longer than that to train. Not to mention that they are designed so that the user has to monitor the whole process, or else they disconnect you. I have no experience with reproducible documents either. Do you have any recommendations on how to do that?

@ductho-le
Would you be so kind as to test my claim about VST's speed? Thank you.
