### 2026-05-23 #10084 @fgdfgfthgr-fox
@rich2355
> It does not have to be a Colab...

Our university doesn't seem to have self‑hosted GPU runner for GitHub action, and even Google Colab Pro+ only give up to 24h run time (which isn't enough for nnUNet). I think the only option left is cloud VM + public notebook, are you okay with that?
