### 2026-05-23 #10084 @rich2355
> [@rich2355](https://github.com/rich2355)
>
> > It does not have to be a Colab...
>
> Our university doesn't seem to have self‑hosted GPU runner for GitHub action, and even Google Colab Pro+ only give up to 24h run time (which isn't enough for nnUNet). I think the only option left is cloud VM + public notebook, are you okay with that?

From my understanding (and correct me if wrong), the high-performance of this package requires the existence of GPU. If so, I can ask the team to decide as asking you to rent a Colab Pro+ for just several small requests may not be ideal for you. Lmk what you think, thanks!
