### 2026-05-23 #10084 @fgdfgfthgr-fox
I have created an account on [Lambda Labs](https://lambda.ai/), however I am still very new to running things on the cloud. To be honest, I find their documentation very confusing. I will ask my computer science supervisor about it next week and discuss how I should approach it.

Or, in the best case, if ductho-le could be so kind to test the claim...
