### 2026-05-26 26d245d
Fixing the issue of instance segmentation result not saving
