### 2026-05-26 8fbc310
Fixing the issue of instance segmentation result not saving (again)
