+++ workflow.py
@@ -218,17 +218,17 @@ if __name__ == "__main__":
 
     parser = argparse.ArgumentParser(description="Deep Learning Workflow")
     # The default values are for the coverage test using the default data
-    parser.add_argument("--workflow_box", nargs='+', choices=["Training", "Validation", "Test", "Predict"], default=["Training", "Validation", "Predict"],
+    parser.add_argument("--workflow_box", nargs='+', choices=["Training", "Validation", "Test", "Predict"], default=["Training"],
                         help="Workflows to enable")
     parser.add_argument("--segmentation_mode", choices=["Semantic", "Instance"], default="Semantic",
                         help="Segmentation Mode")
     parser.add_argument("--train_dataset_path", type=str, default="Datasets/train", help="Train Dataset Path")
-    parser.add_argument("--augmentation_csv_path", type=str, default="Augmentation Parameters Anisotropic.csv",
+    parser.add_argument("--augmentation_csv_path", type=str, default="Augmentation Parameters Isotropic.csv",
                         help="Csv File for Data Augmentation Settings")
-    parser.add_argument("--train_multiplier", type=int, default=128, help="Train Multiplier (Repeats)")
+    parser.add_argument("--train_multiplier", type=int, default=26, help="Train Multiplier (Repeats)")
     parser.add_argument("--batch_size", type=int, default=2, help="Batch Size")
     parser.add_argument("--pairing_samples", default=True, action="store_true", help="Pairing positive and negative samples in a batch")
-    parser.add_argument("--num_epochs", type=int, default=5, help="Number of Epochs") # Five epoch is enough
+    parser.add_argument("--num_epochs", type=int, default=78, help="Number of Epochs") # Five epoch is enough
     parser.add_argument("--enable_tensorboard", default=True, action="store_true", help="Enable TensorBoard Logging")
     parser.add_argument("--enable_unsupervised", action="store_true", help="Enable Unsupervised Pretraining")
     parser.add_argument("--memory_saving_mode", action="store_true", help="Try save some system memory by dataloading on just single core")
@@ -241,7 +241,7 @@ if __name__ == "__main__":
     parser.add_argument("--predict_dataset_path", type=str, default="Datasets/predict", help="Predict Dataset Path")
     parser.add_argument("--read_existing_model", action="store_true", help="Read Existing Model Weight File")
     parser.add_argument("--existing_model_path", type=str, default="", help="Path to Existing Model Weight File")
-    parser.add_argument("--precision", choices=["32", "16-mixed", "bf16-mixed"], default="32", help="Precision")
+    parser.add_argument("--precision", choices=["32", "16-mixed", "bf16-mixed"], default="bf16-mixed", help="Precision")
     parser.add_argument("--save_model_name", type=str, default="example_name",
                         help="File Name for Model Saved, does not include extension")
     parser.add_argument("--save_model_path", type=str, default="trained_model", help="Path to Save the Model Weight")
@@ -250,14 +250,14 @@ if __name__ == "__main__":
     parser.add_argument("--test_key_name", type=str, default=".", help="hdf5 dataset name")
     parser.add_argument("--predict_key_name", type=str, default=".", help="hdf5 dataset name")
     parser.add_argument("--hw_size", type=int, default=144, help="Height and Width of each Training Patch (px)")
-    parser.add_argument("--d_size", type=int, default=40, help="Depth of each Training Patch (px)")
-    parser.add_argument("--predict_hw_size", type=int, default=96, help="Height and Width of each Patch (px) during prediction")
-    parser.add_argument("--predict_depth_size", type=int, default=32, help="Depth of each Patch (px) during prediction")
-    parser.add_argument("--predict_hw_overlap", type=int, default=24,
[…]
