### 2026-05-26 078e9e9
Adjust commands
