### 2026-05-26 adaa55e
Adjust commands (again)
