+++ benchmark_nnunet.py
@@ -12,9 +12,5 @@ if __name__ == "__main__":
     run("nnUNetv2_plan_and_preprocess -d 001 --verify_dataset_integrity", custom_env=my_env)
     run("nnUNetv2_train 001 3d_fullres 0", custom_env=my_env)
     end_time_first_fold = time.time()
-    run("nnUNetv2_train 001 3d_fullres 1", custom_env=my_env)
-    run("nnUNetv2_train 001 3d_fullres 2", custom_env=my_env)
-    run("nnUNetv2_train 001 3d_fullres 3", custom_env=my_env)
-    end_time = time.time()
     print("Total time for a single fold of nnUNet training: ", end_time_first_fold - start_time)
-    print("Total time for nnUNet training: ", end_time - start_time)
+    print("Estimated total time for 4 folds nnUNet training: ", 4 * (end_time_first_fold - start_time))
