### 2026-05-30 #10084 @fgdfgfthgr-fox
Archive in zenodo: https://doi.org/10.5281/zenodo.20454682
I have checked the title, author list and license.
@rich2355
