### 2026-05-31 dc4fe9b
Add missing doi to paper.bib
