+++ paper/paper.bib
@@ -18,7 +18,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={15},
   pages={2424--2426},
   year={2017},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btx180}
 }
 
 @article{berg2019ilastik,
@@ -38,7 +39,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   booktitle={Artificial intelligence in medical imaging: Opportunities, applications and risks},
   pages={39--46},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-319-94878-2_4}
 }
 
 @article{gallusser2022deep,
@@ -49,7 +51,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={2},
   pages={e202208005},
   year={2022},
-  publisher={Rockefeller University Press}
+  publisher={Rockefeller University Press},
+  doi={10.1083/jcb.202208005}
 }
 
 @article{gomez2021deepimagej,
@@ -71,7 +74,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={9},
   pages={677--680},
   year={2018},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41592-018-0106-z}
 }
 
 @article{isensee2021nnu,
@@ -93,7 +97,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={4},
   pages={108035},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jsb.2023.108035}
 }
 
 @article{kamnitsas2017efficient,
@@ -103,7 +108,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   volume={36},
   pages={61--78},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.media.2016.10.004}
 }
 
 @article{li2022auto,
@@ -114,7 +120,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={3},
   pages={e0265567},
   year={2022},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pone.0265567},
 }
 
 @inproceedings{li2017compactness,
@@ -123,7 +130,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   booktitle={International conference on information processing in medical imaging},
   pages={348--360},
   year={2017},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-59050-9_28}
 }
 
 @article{mekuvc2020automatic,
@@ -133,7 +141,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   volume={119},
   pages={103693},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.compbiomed.2020.103693}
 }
 
 @inproceedings{milletari2016v,
@@ -142,7 +151,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
[…]
