+++ paper/paper.bib
@@ -303,7 +303,7 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={6},
   pages={925--927},
   year={2024},
-  publisher={Nature Publishing Group US New York}.
+  publisher={Nature Publishing Group US New York},
   doi={10.1038/s41592-024-02295-6}
 }
 
