### 2026-05-31 #10084 @fgdfgfthgr-fox
I added the missing DOIs. Is that everything? @rich2355
