+++ docs/sourcemd.md
@@ -34,7 +34,7 @@ To contributor: The [GitHub page](https://github.com/fgdfgfthgr-fox/Volume_Seg_T
   - Custom implemented augmentations specialised in volumetric images
   - Process many depth slices all once, rather than slice by slice
   - Support fp16/bf16 training
-  - Up to 20x speedup compare to nnUNet
+  - Comparable to the faster speed of nnUNet
 - Versatile
   - Semantic Segmentation & Instance Segmentation
   - Isotropic & anisotropic images
