+++ paper/paper.bib
@@ -30,7 +30,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={12},
   pages={1226--1232},
   year={2019},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41592-019-0582-9}
 }
 
 @incollection{erickson2019deep,
@@ -63,7 +64,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={10},
   pages={1192--1195},
   year={2021},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41592-021-01262-9}
 }
 
 @article{haberl2018cdeep3m,
@@ -86,7 +88,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={2},
   pages={203--211},
   year={2021},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41592-020-01008-z}
 }
 
 @article{jadav2023beyond,
@@ -163,7 +166,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={7},
   pages={3523--3542},
   year={2021},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPAMI.2021.3059968}
 }
 
 @article{pang2020deep,
@@ -174,7 +178,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={2},
   pages={227--248},
   year={2020},
-  publisher={SAGE Publications Sage CA: Los Angeles, CA}
+  publisher={SAGE Publications Sage CA: Los Angeles, CA},
+  doi={10.3102/1076998619872761}
 }
 
 @article{paszke2019pytorch,
@@ -193,7 +198,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={7},
   pages={671--675},
   year={2012},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/nmeth.2089}
 }
 
 @article{von2021democratising,
@@ -204,7 +210,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   number={1},
   pages={2276},
   year={2021},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41467-021-22518-0}
 }
 
 @article{huang2025generalist,
@@ -271,7 +278,8 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   booktitle={International Conference on Medical image computing and computer-assisted intervention},
   pages={234--241},
   year={2015},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-24574-4_28}
 }
 
 @inproceedings{liu2021swin,
@@ -280,7 +288,7 @@ url={https://www.proquest.com/working-papers/gradio-hassle-free-sharing-testing-
   booktitle={Proceedings of the IEEE/CVF international conference on computer vision},
   pages={10012--10022},
   year={2021},
-  doi={10.1109/iccv48922.2021.00986},
+  doi={10.1109/iccv48922.2021.00986}
 }
 
 @article{franco2025biapy,
