### 2026-06-16 1f0111d
Update lightning requirement to a safer version.
