+++ requirements.txt
@@ -3,7 +3,7 @@ pandas>=2.1.0
 pyimagej>=1.4.1
 scikit-image>=0.21.0
 matplotlib>=3.8.0
-lightning>=2.3.3
+lightning>=2.6.4
 gradio>=6.8.0
 h5py>=3.12.1
 scipy>=1.11.2
