### 2026-06-18 2c01e2d
Add CITATION.cff
