+++ CITATION.cff
@@ -1,0 +1,60 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Huang
+  given-names: Yuyao
+  name-particle: (Daniel)
+  orcid: "https://orcid.org/0009-0007-7171-5029"
+- family-names: Jadav
+  given-names: Nickhil
+  orcid: "https://orcid.org/0009-0009-2754-1097"
+- family-names: Rutter
+  given-names: Georgia
+  orcid: "https://orcid.org/0009-0005-6609-0576"
+- family-names: Bostina
+  given-names: Mihnea
+  orcid: "https://orcid.org/0000-0003-3621-3772"
+- family-names: Harland
+  given-names: Duane
+  orcid: "https://orcid.org/0000-0002-1204-054X"
+- family-names: Szymanski
+  given-names: Lech
+  orcid: "https://orcid.org/0000-0002-5192-0304"
+doi: 10.5281/zenodo.20454682
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Huang
+    given-names: Yuyao
+    name-particle: (Daniel)
+    orcid: "https://orcid.org/0009-0007-7171-5029"
+  - family-names: Jadav
+    given-names: Nickhil
+    orcid: "https://orcid.org/0009-0009-2754-1097"
+  - family-names: Rutter
+    given-names: Georgia
+    orcid: "https://orcid.org/0009-0005-6609-0576"
+  - family-names: Bostina
+    given-names: Mihnea
+    orcid: "https://orcid.org/0000-0003-3621-3772"
+  - family-names: Harland
+    given-names: Duane
+    orcid: "https://orcid.org/0000-0002-1204-054X"
+  - family-names: Szymanski
+    given-names: Lech
+    orcid: "https://orcid.org/0000-0002-5192-0304"
+  date-published: 2026-06-17
+  doi: 10.21105/joss.10084
+  issn: 2475-9066
+  issue: 122
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 10084
+  title: "VST: A Python-based deep learning tool for segmenting electron
+    microscopy samples"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.10084"
+  volume: 11
+title: "VST: A Python-based deep learning tool for segmenting electron
+  microscopy samples"
