### 2026-06-18 1b8c447
Add mention of the new paper in readme.
