### 2026-06-25 03a6dd1
Some memory optimizations for prediction (stage 1)
