### 2026-06-29 b8b7b83
Some memory optimizations for prediction (stage 2)
