+++ workflow.py
@@ -33,6 +33,157 @@ def start_tensorboard(args):
     tensorboard_url = f'http://localhost:{tensorboard_port}'
     webbrowser.open(tensorboard_url)
 
+def generalised_train(TD, UTD, VD, TeD, instance_mode, desired_num_workers, persistent_workers):
+    assert args.train_multiplier != "None", "Received None for train_multiplier! Did you forgot to press the 'Calculate Training Repeats and Epochs' button?'"
+    train_dataset = TD(args.train_dataset_path, args.augmentation_csv_path,
+                       int(args.train_multiplier),
+                       args.hw_size, args.d_size, instance_mode,
+                       args.contour_map_width, args.train_key_name)
+    if args.enable_unsupervised:
+        unsupervised_train_dataset = UTD(args.unsupervised_train_dataset_path,
+                                         args.augmentation_csv_path, int(args.unsupervised_train_multiplier),
+                                         args.hw_size, args.d_size, args.train_key_name)
+    else:
+        unsupervised_train_dataset = None
+    train_dataset = DataComponents.CollectedDataset(train_dataset, unsupervised_train_dataset)
+    sampler = DataComponents.CollectedSampler(train_dataset, args.batch_size, unsupervised_train_dataset)
+    collate_fn = DataComponents.custom_collate
+    train_loader = DataLoader(dataset=train_dataset, batch_size=args.batch_size,
+                              collate_fn=collate_fn, sampler=sampler,
+                              num_workers=desired_num_workers, persistent_workers=persistent_workers)
+                              #num_workers=0, persistent_workers=False)
+    del train_dataset
+    if 'Validation' in args.workflow_box:
+        val_dataset = VD(args.val_dataset_path, args.hw_size, args.d_size, instance_mode,
+                         args.contour_map_width, args.val_key_name)
+        val_loader = DataLoader(dataset=val_dataset, batch_size=1,
+                                #num_workers=12, persistent_workers=True, pin_memory=True)
+                                num_workers=0, pin_memory=True)
+        del val_dataset
+    else:
+        val_loader = None
+    if 'Test' in args.workflow_box:
+        test_dataset = TeD(args.test_dataset_path, args.hw_size, args.d_size, instance_mode,
+                           args.contour_map_width, args.test_key_name)
+        test_loader = DataLoader(dataset=test_dataset, batch_size=args.batch_size)
+        del test_dataset
+    gc.collect()
+    arch_args = ((args.model_patch_size_z, args.model_patch_size_xy, args.model_patch_size_xy), 4,
+                 args.model_depth * args.model_depth_multiplier, instance_mode)
+    if 'Sparsely Labelled' in args.train_dataset_mode:
+        sparse_train = True
+    else:
+        sparse_train = False
+    if args.read_existing_model:
+        model = PLModule.load_from_checkpoint(args.existing_model_path, weights_only=False)
+    else:
+        model = PLModule(arch_args,
+                         'Validation' in args.workflow_box, args.mid_visualization, instance_mode,
+                         sparse_train, 'Sparsely Labelled' in args.val_dataset_mode,
+                         'Sparsely Labelled' in args.test_dataset_mode, args.enable_tensorboard)
+    if 'Training' in args.workflow_box:
+        logger = create_logger(args)
+    else:
+        logger = False
+    if 'Training' in args.workflow_box:
+        '''if 'Validation' in args.workflow_box:
+            to_monitor = 'Val_epoch_dice'
+        else:
+            to_monitor = 'Train_epoch_dice'''
+        callbacks = []
+        '''model_checkpoint = pl.callbacks.ModelCheckpoint(dirpath=f"{args.save_model_path}",
+                                                        filename=f"{args.save_model_name}",
+                                                        mode="max", monitor=to_monitor,
[…]
