+++ pl_module_dit.py
@@ -1,3 +1,4 @@
+import tifffile
 import math
 
 import lightning.pytorch as pl
@@ -231,42 +232,61 @@ class PLModule(pl.LightningModule):
             self.test_metrics.append(result_tuple)
         return {'loss': result_tuple[0]}
 
-    def predict_step(self, batch, batch_idx, dataloader_idx=0):
-        def apply_augmentation(data, i):
-            if i % 2 == 0:
-                # D
-                data = torch.flip(data, [2])
-            if i % 4 <= 1:
-                # H
-                data = torch.flip(data, [3])
-            if i % 8 <= 3:
-                # W
-                data = torch.flip(data, [4])
-            return data
+    @staticmethod
+    def apply_tta_augmentation(data, i):
+        if i % 2 == 0:
+            # D
+            data = torch.flip(data, [2])
+        if i % 4 <= 1:
+            # H
+            data = torch.flip(data, [3])
+        if i % 8 <= 3:
+            # W
+            data = torch.flip(data, [4])
+        return data
 
+    @staticmethod
+    def remove_padding(tensor, hw_overlap, depth_overlap):
+        if hw_overlap or depth_overlap:
+            tensor = tensor[:, :, (depth_overlap if depth_overlap else slice(None)):
+                                   -(depth_overlap if depth_overlap else slice(None)),
+                                   (hw_overlap if hw_overlap else slice(None)):
+                                   -(hw_overlap if hw_overlap else slice(None)),
+                                   (hw_overlap if hw_overlap else slice(None)):
+                                   -(hw_overlap if hw_overlap else slice(None))]
+        return tensor
+
+    def predict_step(self, batch, batch_idx, dataloader_idx=0):
+        batch, hw_overlap, depth_overlap = batch
         TTA_results = []
         for i in range(8):
-            aug_batch = apply_augmentation(batch, i)
+            aug_batch = self.apply_tta_augmentation(batch, i)
             outputs = self.forward(aug_batch)
 
             if isinstance(outputs, tuple):
                 p_outputs = torch.sigmoid(outputs[0]).to(torch.float16)
                 #p_outputs = torch.sigmoid(torch.mean(torch.stack(outputs[0], dim=0), dim=0)).to(torch.float16)
                 c_outputs = torch.sigmoid(outputs[1]).to(torch.float16)
-                p_outputs = apply_augmentation(p_outputs, i)
-                c_outputs = apply_augmentation(c_outputs, i)
+                p_outputs = self.remove_padding(self.apply_tta_augmentation(p_outputs, i), hw_overlap, depth_overlap)
+                c_outputs = self.remove_padding(self.apply_tta_augmentation(c_outputs, i), hw_overlap, depth_overlap)
                 TTA_results.append((p_outputs, c_outputs))
             else:
                 p_outputs = torch.sigmoid(outputs).to(torch.float16)
                 #p_outputs = torch.sigmoid(torch.mean(torch.stack(outputs, dim=0), dim=0)).to(torch.float16)
-                p_outputs = apply_augmentation(p_outputs, i)
+                p_outputs = self.remove_padding(self.apply_tta_augmentation(p_outputs, i), hw_overlap, depth_overlap)
                 TTA_results.append(p_outputs)
 
         if isinstance(TTA_results[0], tuple):
             p, c = zip(*TTA_results)
-            return torch.mean(torch.stack(p, dim=0), dim=0), torch.mean(torch.stack(c, dim=0), dim=0)
+            #return torch.mean(torch.stack(p, dim=0), dim=0), torch.mean(torch.stack(c, dim=0), dim=0)
+            p = torch.mean(torch.stack(p, dim=0), dim=0)
+            c = torch.mean(torch.stack(c, dim=0), dim=0)
[…]

[…]
