### 2026-06-29 6a0a88e
Address the issues arise from the new fp16 intermediate
