### 2026-06-30 bf76d3b
Some memory optimizations for prediction (stage 3)
