+++ Components/MorphologicalFunctions.py
@@ -3,19 +3,20 @@ from numba import njit, prange
 from heapq import heappush, heappop
 
 
-def geodesicreconstructionbyerosion3d(mask, dynamic):
+def geodesic_reconstruction_by_erosion(mask, dynamic):
     result = mask + dynamic
-    mod_if = True
-    print("Geodesic Reconstructing...")
-    while mod_if:
-        mod_if = False
-        result, mod_if = _forward_scan_c6(mask, result, mod_if)
-        result, mod_if = _backward_scan_c6(mask, result, mod_if)
+    changed = True
+    i = 0
+    while changed:
+        changed = False
+        print(f"Iteration {i} of Geodesic Reconstruction...")
+        result, changed = reconstruct_forward_scan(mask, result, changed)
+        result, changed = reconstruct_backward_scan(mask, result, changed)
     return result
 
 
 @njit
-def _forward_scan_c6(mask, result, mod_if):
+def reconstruct_forward_scan(mask, result, changed):
     for z in range(mask.shape[0]):
         for y in range(mask.shape[1]):
             for x in range(mask.shape[2]):
@@ -32,12 +33,12 @@ def _forward_scan_c6(mask, result, mod_if):
                 min_value = max(min_value, mask[z, y, x])
                 if min_value < current_value:
                     result[z, y, x] = min_value
-                    mod_if = True
-    return result, mod_if
+                    changed = True
+    return result, changed
 
 
 @njit
-def _backward_scan_c6(mask, result, mod_if):
+def reconstruct_backward_scan(mask, result, changed):
     for z in range(mask.shape[0] - 1, -1, -1):
         for y in range(mask.shape[1] - 1, -1, -1):
             for x in range(mask.shape[2] - 1, -1, -1):
@@ -54,18 +55,18 @@ def _backward_scan_c6(mask, result, mod_if):
                 min_value = max(min_value, mask[z, y, x])
                 if min_value < current_value:
                     result[z, y, x] = min_value
-                    mod_if = True
-    return result, mod_if
+                    changed = True
+    return result, changed
 
 
-def chamferdistancetransform3duint16(img):
-    result = np.where(img > 0, np.uint16(np.iinfo(np.uint16).max), np.uint16(0))
-    result = _forward_scan_cham_c6(img, result)
-    result = _backward_scan_cham_c6(img, result)
+def chamfer_distancetransform(img, dtype):
+    max_val = np.iinfo(dtype).max
+    result = np.where(img > 0, dtype(max_val), dtype(0))
+    result = cham_forward_scan(img, result)
+    result = cham_backward_scan(img, result)
     return result
 
 # Borgefors 3D chamfer weights (26‑neighbourhood)
-# Define the Borgefors weights and offsets
 offsets = (
     (1, 0, 0, 3),
     (0, 1, 0, 3),
@@ -95,7 +96,7 @@ offsets = (
     (-1, -1, -1, 5))
 
 @njit(nogil=True)
-def _forward_scan_cham_c6(img, result):
+def cham_forward_scan(img, result):
     for z in range(img.shape[0]):
         for y in range(img.shape[1]):
             for x in range(img.shape[2]):
@@ -123,7 +124,7 @@ def _forward_scan_cham_c6(img, result):
 
 
 @njit(nogil=True)
-def _backward_scan_cham_c6(img, result):
+def cham_backward_scan(img, result):
     for z in range(img.shape[0] - 1, -1, -1):
         for y in range(img.shape[1] - 1, -1, -1):
             for x in range(img.shape[2] - 1, -1, -1):
@@ -151,16 +152,16 @@ def _backward_scan_cham_c6(img, result):
 
 
 @njit(parallel=True, nogil=True)
[…]
