### 2026-06-30 f6d60f6
Attempting to fix training issues related to the new fp16 intermediate dtype (unfinished)
