+++ Components/DataComponents.py
@@ -1,1679 +1,0 @@
-import gc
-import re
-import math
-import os
-import torch
-import imageio
-import random
-import json
-import tifffile
-import time
-import h5py
-import multiprocessing
-import mrcfile
-import zarr
-import shutil
-import bisect
-
-import torch.nn.functional as F
-import torchvision.transforms.v2.functional as T_F
-import numpy as np
-import pandas as pd
-import skimage.morphology as morph
-
-from multiprocessing import shared_memory
-from itertools import product
-from pathlib import Path
-from scipy.ndimage import label
-from torchvision.datasets.folder import has_file_allowed_extension
-from . import Augmentations as Aug
-from . import MorphologicalFunctions as Morph
-from .Welford import welford_mean_std
-
-
-device = "cuda" if torch.cuda.is_available() else "cpu"
-
-
-def multiple_loader(path, key):
-    ext = os.path.splitext(path)[1].lower()
-
-    if ext in ['.h5', '.hdf5', '.he5']:
-        with h5py.File(path, 'r') as f:
-            img = np.array(f[key])
-    elif ext in ['.mrc']:
-        with mrcfile.open(path, mode='r') as mrc:
-            img = mrc.data
-    else:  # All other formats
-        img = imageio.v3.imread(path)
-
-    return img
-
-def path_to_array(path, label=False, key='default', norm_strategy='std'):
-    """
-    Transform a path to an image file into a Pytorch tensor. Can support other image format but usually tif is the only one that can store 3d information.
-
-    Args:
-        path (str): Path to the image file.
-        label (bool): If false, the output will be normalized. Default: False.
-        key (str): If trying to load from a hdf5 file, will load the File object with this name.
-        norm_strategy (str): The strategy used to normalize the image (when isn't a label).
-        Can be 'std' to minus by mean then divide by std. Or 'n1tp1' to put it between negative and positive one. Default: 'std'.
-
-    Returns:
-        torch.Tensor: transformed Tensor.
-    """
-    img = multiple_loader(path, key)
-
-    if label:
-        img_max = img.max()
-        if len(np.unique(img)) <= 2:
-            new_dtype = np.bool_
-        elif img_max <= np.iinfo(np.uint8).max:
-            new_dtype = np.uint8
-        elif img_max <= np.iinfo(np.uint16).max:
-            new_dtype = np.uint16
-        else: # I don't think anyone will use more than the maximum value of uint32...
-            new_dtype = np.uint32
-        img = img.astype(new_dtype, copy=False)
-        return img
-    else:
-        if norm_strategy == 'std':
-            # Take less memory than Numpy's Existing method.
-            mean, std = welford_mean_std(img)
-            mean, std = np.array(mean).astype(np.float32), np.array(std).astype(np.float32)
-            # Take less memory than img = (img - mean) / std
-            out_img = np.empty_like(img, dtype=np.float16)
-            for i in range(0, img.shape[0]):
-                slice_img = img[i].astype(np.float32)
-                slice_img -= mean
-                np.divide(slice_img, std, slice_img)
-                out_img[i] = slice_img
-        '''elif norm_strategy == 'n1tp1':
-            img = img.astype(np.float32, copy=False)
-            q1, q99 = np.percentile(img[non_zero], [1, 99])
-            # To 0 and 1
-            img -= q1
-            img /= (q99-q1)
-            np.clip(img, 0, 1, img)
-            # To -1 and 1
-            img -= 0.5
-            img *= 2'''
-        return out_img
-
-def path_to_array_nonorm(path, key='default'):
-    img = multiple_loader(path, key)
-    if len(img.shape) != 3:
-        raise ValueError(f'Only 3D images are supported (Z, Y, X)! {path} seems to be {len(img.shape)}D! instead')
-    mean, std = welford_mean_std(img)
[…]
