### 2026-07-02 e985aba
Fix the iteration counting issue for Geodesic Reconstruction
