+++ Components/Morphology.py
@@ -19,18 +19,31 @@ def geodesic_reconstruction_by_erosion(mask, dynamic):
 
 @njit
 def reconstruct_forward_scan(mask, result, changed):
+    # 26‑connectivity forward pass: neighbours with lower raster indices
     for z in range(mask.shape[0]):
         for y in range(mask.shape[1]):
             for x in range(mask.shape[2]):
                 current_value = result[z, y, x]
                 min_value = current_value
 
-                if x > 0:
-                    min_value = min(min_value, result[z, y, x-1])
-                if y > 0:
-                    min_value = min(min_value, result[z, y-1, x])
-                if z > 0:
-                    min_value = min(min_value, result[z-1, y, x])
+                # All 13 neighbours that are "before" (z,y,x) in raster order
+                for dz in (-1, 0, 1):
+                    nz = z + dz
+                    if nz < 0 or nz >= mask.shape[0]:
+                        continue
+                    for dy in (-1, 0, 1):
+                        ny = y + dy
+                        if ny < 0 or ny >= mask.shape[1]:
+                            continue
+                        for dx in (-1, 0, 1):
+                            if dz == 0 and dy == 0 and dx == 0:
+                                continue
+                            # Condition: nz < z, or (nz == z and ny < y), or (nz == z and ny == y and nx < x)
+                            if dz < 0 or (dz == 0 and dy < 0) or (dz == 0 and dy == 0 and dx < 0):
+                                nx = x + dx
+                                if nx < 0 or nx >= mask.shape[2]:
+                                    continue
+                                min_value = min(min_value, result[nz, ny, nx])
 
                 min_value = max(min_value, mask[z, y, x])
                 if min_value < current_value:
@@ -41,18 +54,31 @@ def reconstruct_forward_scan(mask, result, changed):
 
 @njit
 def reconstruct_backward_scan(mask, result, changed):
+    # 26‑connectivity backward pass: neighbours with higher raster indices
     for z in range(mask.shape[0] - 1, -1, -1):
         for y in range(mask.shape[1] - 1, -1, -1):
             for x in range(mask.shape[2] - 1, -1, -1):
                 current_value = result[z, y, x]
                 min_value = current_value
 
-                if x < mask.shape[2] - 1:
-                    min_value = min(min_value, result[z, y, x+1])
-                if y < mask.shape[1] - 1:
-                    min_value = min(min_value, result[z, y+1, x])
-                if z < mask.shape[0] - 1:
-                    min_value = min(min_value, result[z+1, y, x])
+                # All 13 neighbours that are "after" (z,y,x) in raster order
+                for dz in (-1, 0, 1):
+                    nz = z + dz
+                    if nz < 0 or nz >= mask.shape[0]:
+                        continue
+                    for dy in (-1, 0, 1):
+                        ny = y + dy
+                        if ny < 0 or ny >= mask.shape[1]:
+                            continue
+                        for dx in (-1, 0, 1):
+                            if dz == 0 and dy == 0 and dx == 0:
+                                continue
+                            # Condition: nz > z, or (nz == z and ny > y), or (nz == z and ny == y and nx > x)
+                            if dz > 0 or (dz == 0 and dy > 0) or (dz == 0 and dy == 0 and dx > 0):
+                                nx = x + dx
+                                if nx < 0 or nx >= mask.shape[2]:
+                                    continue
[…]
