### 2026-07-06 8e8aa6d
Clean up documentation, fixes the inverted offload behaviour
