+++ Components/Augmentations.py
@@ -22,7 +22,7 @@ def sim_low_res(tensor, scale=2):
     interpolation.
 
     Args:
-        tensor (torch.Tensor): Input tensor of shape (C, D, H, W) or (C, H, W).
+        tensor (torch.Tensor): Input tensor of shape (C, D, H, W).
         scale (float): Scale factor for down-sampling and up-sampling. Default is 2.
 
     Returns:
@@ -31,24 +31,21 @@ def sim_low_res(tensor, scale=2):
     shape = tensor.shape
     tensor = tensor.unsqueeze(0)
     tensor = F.interpolate(tensor, scale_factor=scale, mode='nearest-exact')
-    if len(shape) == 4:
-        tensor = F.interpolate(tensor, size=[shape[1], shape[2], shape[3]], mode='trilinear')
-    if len(shape) == 3:
-        tensor = F.interpolate(tensor, size=[shape[1], shape[2]], mode='bicubic')
+    tensor = F.interpolate(tensor, size=[shape[1], shape[2], shape[3]], mode='trilinear')
     return tensor.squeeze(0)
 
 
 def adj_gamma(tensor, gamma, gain=1):
     """
-    Adjust gamma correction for a tensor of shape (C, D, H, W).
+    Adjust gamma correction.
 
     Args:
         tensor (torch.Tensor): Input tensor of shape (C, D, H, W).
         gamma (float): Non-negative gamma correction factor.
-        gain (float, optional): Multiplicative gain. Default is 1.
+        gain (float): Multiplicative gain. Default is 1.
 
     Returns:
-        torch.Tensor: Gamma-adjusted 3D tensor with the same shape as the input.
+        torch.Tensor: Gamma-adjusted 3D tensor.
     """
     min, max = tensor.min(), tensor.max()
     if min - max == 0:
@@ -61,14 +58,14 @@ def adj_gamma(tensor, gamma, gain=1):
 
 def adj_contrast(tensor, contrast):
     """
-    Adjust contrast for a tensor of shape (C, D, H, W).
+    Adjust contrast for a tensor.
 
     Args:
         tensor (torch.Tensor): Input tensor of shape (C, D, H, W).
         contrast (float): Contrast adjustment factor.
 
     Returns:
-        torch.Tensor: Contrast-adjusted 3D tensor with the same shape as the input.
+        torch.Tensor: Contrast-adjusted 3D tensor.
     """
     tensor *= contrast
     return tensor
@@ -76,30 +73,30 @@ def adj_contrast(tensor, contrast):
 
 def adj_brightness(tensor, brightness):
     """
-    Adjust brightness for a tensor of shape (C, D, H, W).
+    Adjust brightness for a tensor.
 
     Args:
         tensor (torch.Tensor): Input tensor of shape (C, D, H, W).
         brightness (float): Brightness adjustment factor.
 
     Returns:
-        torch.Tensor: Brightness-adjusted 3D tensor with the same shape as the input.
+        torch.Tensor: Brightness-adjusted 3D tensor.
     """
-    return torch.clamp(brightness + tensor, -4, 4)
+    return torch.clamp(brightness + tensor)
 
 
 def gaussian_blur_3d(tensor, kernel_size=3, sigma=0.8):
     """
-    Apply 3D Gaussian blur to a 3D tensor.
+    Apply 3D Gaussian blur to a tensor.
     High kernel_size could slow down augmentation significantly.
 
     Args:
-        tensor (torch.Tensor): Input 3D tensor of shape {C, D, H, W}.
+        tensor (torch.Tensor): Input tensor of shape {C, D, H, W}.
         kernel_size (int): Size of the Gaussian kernel. Default: 3.
[…]
