### 2026-07-06 28af32f
Attempt to reduce training memory use
