### 2026-07-06 da6a48b
Improve speed and reduce memory use when generating contour maps.
