### 2026-07-08 6cb9b95
Improve speed and reduce memory use when generating contour maps (again).
