### 2026-07-08 747c1a8
Use a custom function to remove small labels for faster speed and less memory (instance segmentation).
