### 2026-07-11 e850243
Update pytest file to match with the new label size filtering method
