### 2026-07-31 58662c1
Use numba-morph for some morphology functions
