### 2026-07-31 8ad3a0f
Remove old morphology code
