+++ Components/Morphology.py
@@ -4,272 +4,6 @@ from numba.typed import List
 from heapq import heappush, heappop
 
 
-def geodesic_reconstruction_by_erosion(mask, dynamic):
-    result = mask + dynamic
-    changed = True
-    i = 0
-    while changed:
-        changed = False
-        print(f"Iteration {i} of Geodesic Reconstruction...")
-        result, changed = reconstruct_forward_scan(mask, result, changed)
-        result, changed = reconstruct_backward_scan(mask, result, changed)
-        i += 1
-    return result
-
-
-@njit
-def reconstruct_forward_scan(mask, result, changed):
-    # 26‑connectivity forward pass: neighbours with lower raster indices
-    for z in range(mask.shape[0]):
-        for y in range(mask.shape[1]):
-            for x in range(mask.shape[2]):
-                current_value = result[z, y, x]
-                min_value = current_value
-
-                # All 13 neighbours that are "before" (z,y,x) in raster order
-                for dz in (-1, 0, 1):
-                    nz = z + dz
-                    if nz < 0 or nz >= mask.shape[0]:
-                        continue
-                    for dy in (-1, 0, 1):
-                        ny = y + dy
-                        if ny < 0 or ny >= mask.shape[1]:
-                            continue
-                        for dx in (-1, 0, 1):
-                            if dz == 0 and dy == 0 and dx == 0:
-                                continue
-                            # Condition: nz < z, or (nz == z and ny < y), or (nz == z and ny == y and nx < x)
-                            if dz < 0 or (dz == 0 and dy < 0) or (dz == 0 and dy == 0 and dx < 0):
-                                nx = x + dx
-                                if nx < 0 or nx >= mask.shape[2]:
-                                    continue
-                                min_value = min(min_value, result[nz, ny, nx])
-
-                min_value = max(min_value, mask[z, y, x])
-                if min_value < current_value:
-                    result[z, y, x] = min_value
-                    changed = True
-    return result, changed
-
-
-@njit
-def reconstruct_backward_scan(mask, result, changed):
-    # 26‑connectivity backward pass: neighbours with higher raster indices
-    for z in range(mask.shape[0] - 1, -1, -1):
-        for y in range(mask.shape[1] - 1, -1, -1):
-            for x in range(mask.shape[2] - 1, -1, -1):
-                current_value = result[z, y, x]
-                min_value = current_value
-
-                # All 13 neighbours that are "after" (z,y,x) in raster order
-                for dz in (-1, 0, 1):
-                    nz = z + dz
-                    if nz < 0 or nz >= mask.shape[0]:
-                        continue
-                    for dy in (-1, 0, 1):
-                        ny = y + dy
-                        if ny < 0 or ny >= mask.shape[1]:
-                            continue
-                        for dx in (-1, 0, 1):
-                            if dz == 0 and dy == 0 and dx == 0:
-                                continue
-                            # Condition: nz > z, or (nz == z and ny > y), or (nz == z and ny == y and nx > x)
-                            if dz > 0 or (dz == 0 and dy > 0) or (dz == 0 and dy == 0 and dx > 0):
-                                nx = x + dx
-                                if nx < 0 or nx >= mask.shape[2]:
-                                    continue
-                                min_value = min(min_value, result[nz, ny, nx])
-
-                min_value = max(min_value, mask[z, y, x])
[…]
